@@ -1,75 +1,103 @@
 package singleton
 
 import (
-	"sort"
-	"sync"
+	"cmp"
+	"slices"
 
-	"github.com/naiba/nezha/model"
+	"github.com/nezhahq/nezha/model"
+	"github.com/nezhahq/nezha/pkg/utils"
 )
 
-var (
-	ServerList        map[uint64]*model.Server // [ServerID] -> model.Server
-	SecretToID        map[string]uint64        // [ServerSecret] -> ServerID
-	ServerTagToIDList map[string][]uint64      // [ServerTag] -> ServerID
-	ServerLock        sync.RWMutex
+type ServerClass struct {
+	class[uint64, *model.Server]
 
-	SortedServerList         []*model.Server // 用于存储服务器列表的 slice，按照服务器 ID 排序
-	SortedServerListForGuest []*model.Server
-	SortedServerLock         sync.RWMutex
-)
+	uuidToID map[string]uint64
 
-// InitServer 初始化 ServerID <-> Secret 的映射
-func InitServer() {
-	ServerList = make(map[uint64]*model.Server)
-	SecretToID = make(map[string]uint64)
-	ServerTagToIDList = make(map[string][]uint64)
+	sortedListForGuest []*model.Server
 }
 
-// loadServers 加载服务器列表并根据ID排序
-func loadServers() {
-	InitServer()
+func NewServerClass() *ServerClass {
+	sc := &ServerClass{
+		class: class[uint64, *model.Server]{
+			list: make(map[uint64]*model.Server),
+		},
+		uuidToID: make(map[string]uint64),
+	}
+
 	var servers []model.Server
 	DB.Find(&servers)
 	for _, s := range servers {
 		innerS := s
-		innerS.Host = &model.Host{}
-		innerS.State = &model.HostState{}
-		innerS.TaskCloseLock = new(sync.Mutex)
-		ServerList[innerS.ID] = &innerS
-		SecretToID[innerS.Secret] = innerS.ID
-		ServerTagToIDList[innerS.Tag] = append(ServerTagToIDList[innerS.Tag], innerS.ID)
+		model.InitServer(&innerS)
+		sc.list[innerS.ID] = &innerS
+		sc.uuidToID[innerS.UUID] = innerS.ID
 	}
-	ReSortServer()
+	sc.sortList()
+
+	return sc
 }
 
-// ReSortServer 根据服务器ID 对服务器列表进行排序（ID越大越靠前）
-func ReSortServer() {
-	ServerLock.RLock()
-	defer ServerLock.RUnlock()
-	SortedServerLock.Lock()
-	defer SortedServerLock.Unlock()
+func (c *ServerClass) Update(s *model.Server, uuid string) {
+	c.listMu.Lock()
 
-	SortedServerList = []*model.Server{}
-	SortedServerListForGuest = []*model.Server{}
-	for _, s := range ServerList {
-		SortedServerList = append(SortedServerList, s)
-		if !s.HideForGuest {
-			SortedServerListForGuest = append(SortedServerListForGuest, s)
-		}
+	c.list[s.ID] = s
+	if uuid != "" {
+		c.uuidToID[uuid] = s.ID
 	}
 
+	c.listMu.Unlock()
+
+	c.sortList()
+}
+
+func (c *ServerClass) Delete(idList []uint64) {
+	c.listMu.Lock()
+
+	for _, id := range idList {
+		serverUUID := c.list[id].UUID
+		delete(c.uuidToID, serverUUID)
+		delete(c.list, id)
+	}
+
+	c.listMu.Unlock()
+
+	c.sortList()
+}
+
+func (c *ServerClass) GetSortedListForGuest() []*model.Server {
+	c.sortedListMu.RLock()
+	defer c.sortedListMu.RUnlock()
+
+	return slices.Clone(c.sortedListForGuest)
+}
+
+func (c *ServerClass) UUIDToID(uuid string) (id uint64, ok bool) {
+	c.listMu.RLock()
+	defer c.listMu.RUnlock()
+
+	id, ok = c.uuidToID[uuid]
+	return
+}
+
+func (c *ServerClass) sortList() {
+	c.listMu.RLock()
+	defer c.listMu.RUnlock()
+	c.sortedListMu.Lock()
+	defer c.sortedListMu.Unlock()
+
+	c.sortedList = utils.MapValuesToSlice(c.list)
 	// 按照服务器 ID 排序的具体实现（ID越大越靠前）
-	sort.SliceStable(SortedServerList, func(i, j int) bool {
-		if SortedServerList[i].DisplayIndex == SortedServerList[j].DisplayIndex {
-			return SortedServerList[i].ID < SortedServerList[j].ID
+	slices.SortStableFunc(c.sortedList, func(a, b *model.Server) int {
+		if a.DisplayIndex == b.DisplayIndex {
+			return cmp.Compare(a.ID, b.ID)
 		}
-		return SortedServerList[i].DisplayIndex > SortedServerList[j].DisplayIndex
+		return cmp.Compare(b.DisplayIndex, a.DisplayIndex)
 	})
 
-	sort.SliceStable(SortedServerListForGuest, func(i, j int) bool {
-		if SortedServerListForGuest[i].DisplayIndex == SortedServerListForGuest[j].DisplayIndex {
-			return SortedServerListForGuest[i].ID < SortedServerListForGuest[j].ID
+	c.sortedListForGuest = make([]*model.Server, 0, len(c.sortedList))
+	for _, s := range c.sortedList {
+		if !s.HideForGuest {
+			c.sortedListForGuest = append(c.sortedListForGuest, s)
 		}
-		return SortedServerListForGuest[i].DisplayIndex > SortedServerListForGuest[j].DisplayIndex
-	})
+	}
 }
