@@ -1,117 +1,139 @@
 package singleton
 
 import (
+	"cmp"
 	"fmt"
+	"iter"
 	"log"
-	"sort"
+	"maps"
+	"slices"
 	"strings"
 	"sync"
 	"time"
 
-	"github.com/nicksnyder/go-i18n/v2/i18n"
-
-	"github.com/naiba/nezha/model"
-	pb "github.com/naiba/nezha/proto"
+	"github.com/gin-gonic/gin"
+	"github.com/jinzhu/copier"
+	"github.com/nezhahq/nezha/model"
+	"github.com/nezhahq/nezha/pkg/utils"
+	pb "github.com/nezhahq/nezha/proto"
+	"golang.org/x/exp/constraints"
 )
 
 const (
 	_CurrentStatusSize = 30 // 统计 15 分钟内的数据为当前状态
 )
 
-var ServiceSentinelShared *ServiceSentinel
+type serviceResponseItem struct {
+	model.ServiceResponseItem
+
+	service *model.Service
+}
 
 type ReportData struct {
 	Data     *pb.TaskResult
 	Reporter uint64
 }
 
-// _TodayStatsOfMonitor 今日监控记录
-type _TodayStatsOfMonitor struct {
+// _TodayStatsOfService 今日监控记录
+type _TodayStatsOfService struct {
 	Up    int     // 今日在线计数
 	Down  int     // 今日离线计数
 	Delay float32 // 今日平均延迟
-}
-
-// NewServiceSentinel 创建服务监控器
-func NewServiceSentinel(serviceSentinelDispatchBus chan<- model.Monitor) {
-	ServiceSentinelShared = &ServiceSentinel{
-		serviceReportChannel:                    make(chan ReportData, 200),
-		serviceStatusToday:                      make(map[uint64]*_TodayStatsOfMonitor),
-		serviceCurrentStatusIndex:               make(map[uint64]*indexStore),
-		serviceCurrentStatusData:                make(map[uint64][]*pb.TaskResult),
-		lastStatus:                              make(map[uint64]int),
-		serviceResponseDataStoreCurrentUp:       make(map[uint64]uint64),
-		serviceResponseDataStoreCurrentDown:     make(map[uint64]uint64),
-		serviceResponseDataStoreCurrentAvgDelay: make(map[uint64]float32),
-		serviceResponsePing:                     make(map[uint64]map[uint64]*pingStore),
-		monitors:                                make(map[uint64]*model.Monitor),
-		sslCertCache:                            make(map[uint64]string),
-		// 30天数据缓存
-		monthlyStatus: make(map[uint64]*model.ServiceItemResponse),
-		dispatchBus:   serviceSentinelDispatchBus,
-	}
-	// 加载历史记录
-	ServiceSentinelShared.loadMonitorHistory()
-
-	year, month, day := time.Now().Date()
-	today := time.Date(year, month, day, 0, 0, 0, 0, Loc)
-
-	var mhs []model.MonitorHistory
-	// 加载当日记录
-	DB.Where("created_at >= ?", today).Find(&mhs)
-	totalDelay := make(map[uint64]float32)
-	totalDelayCount := make(map[uint64]float32)
-	for i := 0; i < len(mhs); i++ {
-		totalDelay[mhs[i].MonitorID] += mhs[i].AvgDelay
-		totalDelayCount[mhs[i].MonitorID]++
-		ServiceSentinelShared.serviceStatusToday[mhs[i].MonitorID].Up += int(mhs[i].Up)
-		ServiceSentinelShared.monthlyStatus[mhs[i].MonitorID].TotalUp += mhs[i].Up
-		ServiceSentinelShared.serviceStatusToday[mhs[i].MonitorID].Down += int(mhs[i].Down)
-		ServiceSentinelShared.monthlyStatus[mhs[i].MonitorID].TotalDown += mhs[i].Down
-	}
-	for id, delay := range totalDelay {
-		ServiceSentinelShared.serviceStatusToday[id].Delay = delay / float32(totalDelayCount[id])
-	}
-
-	// 启动服务监控器
-	go ServiceSentinelShared.worker()
-
-	// 每日将游标往后推一天
-	_, err := Cron.AddFunc("0 0 0 * * *", ServiceSentinelShared.refreshMonthlyServiceStatus)
-	if err != nil {
-		panic(err)
-	}
 }
 
 /*
 使用缓存 channel，处理上报的 Service 请求结果，然后判断是否需要报警
 需要记录上一次的状态信息
 
-加锁顺序：serviceResponseDataStoreLock > monthlyStatusLock > monitorsLock
+加锁顺序：serviceResponseDataStoreLock > monthlyStatusLock > servicesLock
 */
 type ServiceSentinel struct {
 	// 服务监控任务上报通道
 	serviceReportChannel chan ReportData // 服务状态汇报管道
 	// 服务监控任务调度通道
-	dispatchBus chan<- model.Monitor
+	dispatchBus chan<- *model.Service
 
 	serviceResponseDataStoreLock            sync.RWMutex
-	serviceStatusToday                      map[uint64]*_TodayStatsOfMonitor // [monitor_id] -> _TodayStatsOfMonitor
-	serviceCurrentStatusIndex               map[uint64]*indexStore           // [monitor_id] -> 该监控ID对应的 serviceCurrentStatusData 的最新索引下标
-	serviceCurrentStatusData                map[uint64][]*pb.TaskResult      // [monitor_id] -> []model.MonitorHistory
-	serviceResponseDataStoreCurrentUp       map[uint64]uint64                // [monitor_id] -> 当前服务在线计数
-	serviceResponseDataStoreCurrentDown     map[uint64]uint64                // [monitor_id] -> 当前服务离线计数
-	serviceResponseDataStoreCurrentAvgDelay map[uint64]float32               // [monitor_id] -> 当前服务离线计数
-	serviceResponsePing                     map[uint64]map[uint64]*pingStore // [monitor_id] -> ClientID -> delay
-	lastStatus                              map[uint64]int
-	sslCertCache                            map[uint64]string
+	serviceStatusToday                      map[uint64]*_TodayStatsOfService // [service_id] -> _TodayStatsOfService
+	serviceCurrentStatusIndex               map[uint64]*indexStore           // [service_id] -> 该监控ID对应的 serviceCurrentStatusData 的最新索引下标
+	serviceCurrentStatusData                map[uint64][]*pb.TaskResult      // [service_id] -> []model.ServiceHistory
+	serviceResponseDataStoreCurrentUp       map[uint64]uint64                // [service_id] -> 当前服务在线计数
+	serviceResponseDataStoreCurrentDown     map[uint64]uint64                // [service_id] -> 当前服务离线计数
+	serviceResponseDataStoreCurrentAvgDelay map[uint64]float32               // [service_id] -> 当前服务离线计数
+	serviceResponsePing                     map[uint64]map[uint64]*pingStore // [service_id] -> ClientID -> delay
+	lastStatus                              map[uint64]uint8
+	tlsCertCache                            map[uint64]string
 
-	monitorsLock sync.RWMutex
-	monitors     map[uint64]*model.Monitor // [monitor_id] -> model.Monitor
+	servicesLock    sync.RWMutex
+	serviceListLock sync.RWMutex
+	services        map[uint64]*model.Service
+	serviceList     []*model.Service
 
 	// 30天数据缓存
 	monthlyStatusLock sync.Mutex
-	monthlyStatus     map[uint64]*model.ServiceItemResponse // [monitor_id] -> model.ServiceItemResponse
+	monthlyStatus     map[uint64]*serviceResponseItem
+
+	// references
+	serverc       *ServerClass
+	notificationc *NotificationClass
+	crc           *CronClass
+}
+
+// NewServiceSentinel 创建服务监控器
+func NewServiceSentinel(serviceSentinelDispatchBus chan<- *model.Service, sc *ServerClass, nc *NotificationClass, crc *CronClass) (*ServiceSentinel, error) {
+	ss := &ServiceSentinel{
+		serviceReportChannel:                    make(chan ReportData, 200),
+		serviceStatusToday:                      make(map[uint64]*_TodayStatsOfService),
+		serviceCurrentStatusIndex:               make(map[uint64]*indexStore),
+		serviceCurrentStatusData:                make(map[uint64][]*pb.TaskResult),
+		lastStatus:                              make(map[uint64]uint8),
+		serviceResponseDataStoreCurrentUp:       make(map[uint64]uint64),
+		serviceResponseDataStoreCurrentDown:     make(map[uint64]uint64),
+		serviceResponseDataStoreCurrentAvgDelay: make(map[uint64]float32),
+		serviceResponsePing:                     make(map[uint64]map[uint64]*pingStore),
+		services:                                make(map[uint64]*model.Service),
+		tlsCertCache:                            make(map[uint64]string),
+		// 30天数据缓存
+		monthlyStatus: make(map[uint64]*serviceResponseItem),
+		dispatchBus:   serviceSentinelDispatchBus,
+
+		serverc:       sc,
+		notificationc: nc,
+		crc:           crc,
+	}
+	// 加载历史记录
+	ss.loadServiceHistory()
+
+	year, month, day := time.Now().Date()
+	today := time.Date(year, month, day, 0, 0, 0, 0, Loc)
+
+	var mhs []model.ServiceHistory
+	// 加载当日记录
+	DB.Where("created_at >= ?", today).Find(&mhs)
+	totalDelay := make(map[uint64]float32)
+	totalDelayCount := make(map[uint64]float32)
+	for i := 0; i < len(mhs); i++ {
+		totalDelay[mhs[i].ServiceID] += mhs[i].AvgDelay
+		totalDelayCount[mhs[i].ServiceID]++
+		ss.serviceStatusToday[mhs[i].ServiceID].Up += int(mhs[i].Up)
+		ss.monthlyStatus[mhs[i].ServiceID].TotalUp += mhs[i].Up
+		ss.serviceStatusToday[mhs[i].ServiceID].Down += int(mhs[i].Down)
+		ss.monthlyStatus[mhs[i].ServiceID].TotalDown += mhs[i].Down
+	}
+	for id, delay := range totalDelay {
+		ss.serviceStatusToday[id].Delay = delay / float32(totalDelayCount[id])
+	}
+
+	// 启动服务监控器
+	go ss.worker()
+
+	// 每日将游标往后推一天
+	_, err := crc.AddFunc("0 0 0 * * *", ss.refreshMonthlyServiceStatus)
+	if err != nil {
+		return nil, err
+	}
+
+	return ss, nil
 }
 
 type indexStore struct {
@@ -159,67 +181,58 @@ func (ss *ServiceSentinel) Dispatch(r ReportData) {
 	ss.serviceReportChannel <- r
 }
 
-func (ss *ServiceSentinel) Monitors() []*model.Monitor {
-	ss.monitorsLock.RLock()
-	defer ss.monitorsLock.RUnlock()
-	var monitors []*model.Monitor
-	for _, v := range ss.monitors {
-		monitors = append(monitors, v)
-	}
-	sort.SliceStable(monitors, func(i, j int) bool {
-		return monitors[i].ID < monitors[j].ID
+func (ss *ServiceSentinel) UpdateServiceList() {
+	ss.servicesLock.RLock()
+	defer ss.servicesLock.RUnlock()
+
+	ss.serviceListLock.Lock()
+	defer ss.serviceListLock.Unlock()
+
+	ss.serviceList = utils.MapValuesToSlice(ss.services)
+	slices.SortFunc(ss.serviceList, func(a, b *model.Service) int {
+		return cmp.Compare(a.ID, b.ID)
 	})
-	return monitors
 }
 
-// loadMonitorHistory 加载服务监控器的历史状态信息
-func (ss *ServiceSentinel) loadMonitorHistory() {
-	var monitors []*model.Monitor
-	err := DB.Find(&monitors).Error
+// loadServiceHistory 加载服务监控器的历史状态信息
+func (ss *ServiceSentinel) loadServiceHistory() {
+	var services []*model.Service
+	err := DB.Find(&services).Error
 	if err != nil {
 		panic(err)
 	}
 
-	ss.serviceResponseDataStoreLock.Lock()
-	defer ss.serviceResponseDataStoreLock.Unlock()
-	ss.monthlyStatusLock.Lock()
-	defer ss.monthlyStatusLock.Unlock()
-	ss.monitorsLock.Lock()
-	defer ss.monitorsLock.Unlock()
-
-	for i := 0; i < len(monitors); i++ {
-		// 旧版本可能不存在通知组 为其设置默认组
-		if monitors[i].NotificationTag == "" {
-			monitors[i].NotificationTag = "default"
-			DB.Save(monitors[i])
-		}
-		task := *monitors[i]
+	for i := 0; i < len(services); i++ {
+		task := services[i]
 		// 通过cron定时将服务监控任务传递给任务调度管道
-		monitors[i].CronJobID, err = Cron.AddFunc(task.CronSpec(), func() {
+		services[i].CronJobID, err = ss.crc.AddFunc(task.CronSpec(), func() {
 			ss.dispatchBus <- task
 		})
 		if err != nil {
 			panic(err)
 		}
-		ss.monitors[monitors[i].ID] = monitors[i]
-		ss.serviceCurrentStatusData[monitors[i].ID] = make([]*pb.TaskResult, _CurrentStatusSize)
-		ss.serviceStatusToday[monitors[i].ID] = &_TodayStatsOfMonitor{}
+		ss.services[services[i].ID] = services[i]
+		ss.serviceCurrentStatusData[services[i].ID] = make([]*pb.TaskResult, _CurrentStatusSize)
+		ss.serviceStatusToday[services[i].ID] = &_TodayStatsOfService{}
 	}
+	ss.serviceList = services
 
 	year, month, day := time.Now().Date()
 	today := time.Date(year, month, day, 0, 0, 0, 0, Loc)
 
-	for i := 0; i < len(monitors); i++ {
-		ServiceSentinelShared.monthlyStatus[monitors[i].ID] = &model.ServiceItemResponse{
-			Monitor: monitors[i],
-			Delay:   &[30]float32{0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0},
-			Up:      &[30]int{0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0},
-			Down:    &[30]int{0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0},
+	for i := 0; i < len(services); i++ {
+		ss.monthlyStatus[services[i].ID] = &serviceResponseItem{
+			service: services[i],
+			ServiceResponseItem: model.ServiceResponseItem{
+				Delay: &[30]float32{0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0},
+				Up:    &[30]int{0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0},
+				Down:  &[30]int{0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0},
+			},
 		}
 	}
 
 	// 加载服务监控历史记录
-	var mhs []model.MonitorHistory
+	var mhs []model.ServiceHistory
 	DB.Where("created_at > ? AND created_at < ?", today.AddDate(0, 0, -29), today).Find(&mhs)
 	var delayCount = make(map[int]int)
 	for i := 0; i < len(mhs); i++ {
@@ -227,83 +240,89 @@ func (ss *ServiceSentinel) loadMonitorHistory() {
 		if dayIndex < 0 {
 			continue
 		}
-		ServiceSentinelShared.monthlyStatus[mhs[i].MonitorID].Delay[dayIndex] = (ServiceSentinelShared.monthlyStatus[mhs[i].MonitorID].Delay[dayIndex]*float32(delayCount[dayIndex]) + mhs[i].AvgDelay) / float32(delayCount[dayIndex]+1)
+		ss.monthlyStatus[mhs[i].ServiceID].Delay[dayIndex] = (ss.monthlyStatus[mhs[i].ServiceID].Delay[dayIndex]*float32(delayCount[dayIndex]) + mhs[i].AvgDelay) / float32(delayCount[dayIndex]+1)
 		delayCount[dayIndex]++
-		ServiceSentinelShared.monthlyStatus[mhs[i].MonitorID].Up[dayIndex] += int(mhs[i].Up)
-		ServiceSentinelShared.monthlyStatus[mhs[i].MonitorID].TotalUp += mhs[i].Up
-		ServiceSentinelShared.monthlyStatus[mhs[i].MonitorID].Down[dayIndex] += int(mhs[i].Down)
-		ServiceSentinelShared.monthlyStatus[mhs[i].MonitorID].TotalDown += mhs[i].Down
+		ss.monthlyStatus[mhs[i].ServiceID].Up[dayIndex] += int(mhs[i].Up)
+		ss.monthlyStatus[mhs[i].ServiceID].TotalUp += mhs[i].Up
+		ss.monthlyStatus[mhs[i].ServiceID].Down[dayIndex] += int(mhs[i].Down)
+		ss.monthlyStatus[mhs[i].ServiceID].TotalDown += mhs[i].Down
 	}
 }
 
-func (ss *ServiceSentinel) OnMonitorUpdate(m model.Monitor) error {
+func (ss *ServiceSentinel) Update(m *model.Service) error {
 	ss.serviceResponseDataStoreLock.Lock()
 	defer ss.serviceResponseDataStoreLock.Unlock()
 	ss.monthlyStatusLock.Lock()
 	defer ss.monthlyStatusLock.Unlock()
-	ss.monitorsLock.Lock()
-	defer ss.monitorsLock.Unlock()
+	ss.servicesLock.Lock()
+	defer ss.servicesLock.Unlock()
 
 	var err error
 	// 写入新任务
-	m.CronJobID, err = Cron.AddFunc(m.CronSpec(), func() {
+	m.CronJobID, err = ss.crc.AddFunc(m.CronSpec(), func() {
 		ss.dispatchBus <- m
 	})
 	if err != nil {
 		return err
 	}
-	if ss.monitors[m.ID] != nil {
+	if ss.services[m.ID] != nil {
 		// 停掉旧任务
-		Cron.Remove(ss.monitors[m.ID].CronJobID)
+		ss.crc.Remove(ss.services[m.ID].CronJobID)
 	} else {
 		// 新任务初始化数据
-		ss.monthlyStatus[m.ID] = &model.ServiceItemResponse{
-			Monitor: &m,
-			Delay:   &[30]float32{0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0},
-			Up:      &[30]int{0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0},
-			Down:    &[30]int{0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0},
+		ss.monthlyStatus[m.ID] = &serviceResponseItem{
+			service: m,
+			ServiceResponseItem: model.ServiceResponseItem{
+				Delay: &[30]float32{0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0},
+				Up:    &[30]int{0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0},
+				Down:  &[30]int{0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0},
+			},
 		}
 		ss.serviceCurrentStatusData[m.ID] = make([]*pb.TaskResult, _CurrentStatusSize)
-		ss.serviceStatusToday[m.ID] = &_TodayStatsOfMonitor{}
+		ss.serviceStatusToday[m.ID] = &_TodayStatsOfService{}
 	}
 	// 更新这个任务
-	ss.monitors[m.ID] = &m
+	ss.services[m.ID] = m
 	return nil
 }
 
-func (ss *ServiceSentinel) OnMonitorDelete(id uint64) {
+func (ss *ServiceSentinel) Delete(ids []uint64) {
 	ss.serviceResponseDataStoreLock.Lock()
 	defer ss.serviceResponseDataStoreLock.Unlock()
 	ss.monthlyStatusLock.Lock()
 	defer ss.monthlyStatusLock.Unlock()
-	ss.monitorsLock.Lock()
-	defer ss.monitorsLock.Unlock()
+	ss.servicesLock.Lock()
+	defer ss.servicesLock.Unlock()
 
-	delete(ss.serviceCurrentStatusIndex, id)
-	delete(ss.serviceCurrentStatusData, id)
-	delete(ss.lastStatus, id)
-	delete(ss.serviceResponseDataStoreCurrentUp, id)
-	delete(ss.serviceResponseDataStoreCurrentDown, id)
-	delete(ss.serviceResponseDataStoreCurrentAvgDelay, id)
-	delete(ss.sslCertCache, id)
-	delete(ss.serviceStatusToday, id)
+	for _, id := range ids {
+		delete(ss.serviceCurrentStatusIndex, id)
+		delete(ss.serviceCurrentStatusData, id)
+		delete(ss.lastStatus, id)
+		delete(ss.serviceResponseDataStoreCurrentUp, id)
+		delete(ss.serviceResponseDataStoreCurrentDown, id)
+		delete(ss.serviceResponseDataStoreCurrentAvgDelay, id)
+		delete(ss.tlsCertCache, id)
+		delete(ss.serviceStatusToday, id)
 
-	// 停掉定时任务
-	Cron.Remove(ss.monitors[id].CronJobID)
-	delete(ss.monitors, id)
+		// 停掉定时任务
+		ss.crc.Remove(ss.services[id].CronJobID)
+		delete(ss.services, id)
 
-	delete(ss.monthlyStatus, id)
+		delete(ss.monthlyStatus, id)
+	}
 }
 
-func (ss *ServiceSentinel) LoadStats() map[uint64]*model.ServiceItemResponse {
+func (ss *ServiceSentinel) LoadStats() map[uint64]*serviceResponseItem {
+	ss.servicesLock.RLock()
+	defer ss.servicesLock.RUnlock()
 	ss.serviceResponseDataStoreLock.RLock()
 	defer ss.serviceResponseDataStoreLock.RUnlock()
 	ss.monthlyStatusLock.Lock()
 	defer ss.monthlyStatusLock.Unlock()
 
 	// 刷新最新一天的数据
-	for k := range ss.monitors {
-		ss.monthlyStatus[k].Monitor = ss.monitors[k]
+	for k := range ss.services {
+		ss.monthlyStatus[k].service = ss.services[k]
 		v := ss.serviceStatusToday[k]
 
 		// 30 天在线率，
@@ -319,7 +338,7 @@ func (ss *ServiceSentinel) LoadStats() map[uint64]*model.ServiceItemResponse {
 		ss.monthlyStatus[k].Delay[29] = v.Delay
 	}
 
-	// 最后 5 分钟的状态 与 monitor 对象填充
+	// 最后 5 分钟的状态 与 service 对象填充
 	for k, v := range ss.serviceResponseDataStoreCurrentDown {
 		ss.monthlyStatus[k].CurrentDown = v
 	}
@@ -330,42 +349,95 @@ func (ss *ServiceSentinel) LoadStats() map[uint64]*model.ServiceItemResponse {
 	return ss.monthlyStatus
 }
 
+func (ss *ServiceSentinel) CopyStats() map[uint64]model.ServiceResponseItem {
+	var stats map[uint64]*serviceResponseItem
+	copier.Copy(&stats, ss.LoadStats())
+
+	sri := make(map[uint64]model.ServiceResponseItem)
+	for k, service := range stats {
+		if !service.service.EnableShowInService {
+			delete(stats, k)
+			continue
+		}
+
+		service.ServiceName = service.service.Name
+		sri[k] = service.ServiceResponseItem
+	}
+
+	return sri
+}
+
+func (ss *ServiceSentinel) Get(id uint64) (s *model.Service, ok bool) {
+	ss.servicesLock.RLock()
+	defer ss.servicesLock.RUnlock()
+
+	s, ok = ss.services[id]
+	return
+}
+
+func (ss *ServiceSentinel) GetList() map[uint64]*model.Service {
+	ss.servicesLock.RLock()
+	defer ss.servicesLock.RUnlock()
+
+	return maps.Clone(ss.services)
+}
+
+func (ss *ServiceSentinel) GetSortedList() []*model.Service {
+	ss.serviceListLock.RLock()
+	defer ss.serviceListLock.RUnlock()
+
+	return slices.Clone(ss.serviceList)
+}
+
+func (ss *ServiceSentinel) CheckPermission(c *gin.Context, idList iter.Seq[uint64]) bool {
+	ss.servicesLock.RLock()
+	defer ss.servicesLock.RUnlock()
+
+	for id := range idList {
+		if s, ok := ss.services[id]; ok {
+			if !s.HasPermission(c) {
+				return false
+			}
+		}
+	}
+	return true
+}
+
 // worker 服务监控的实际工作流程
 func (ss *ServiceSentinel) worker() {
 	// 从服务状态汇报管道获取汇报的服务数据
 	for r := range ss.serviceReportChannel {
-		if ss.monitors[r.Data.GetId()] == nil || ss.monitors[r.Data.GetId()].ID == 0 {
-			log.Printf("NEZHA>> 错误的服务监控上报 %+v", r)
+		css, _ := ss.Get(r.Data.GetId())
+		if css == nil || css.ID == 0 {
+			log.Printf("NEZHA>> Incorrect service monitor report %+v", r)
 			continue
 		}
+		css = nil
 		mh := r.Data
 		if mh.Type == model.TaskTypeTCPPing || mh.Type == model.TaskTypeICMPPing {
-			monitorTcpMap, ok := ss.serviceResponsePing[mh.GetId()]
+			serviceTcpMap, ok := ss.serviceResponsePing[mh.GetId()]
 			if !ok {
-				monitorTcpMap = make(map[uint64]*pingStore)
-				ss.serviceResponsePing[mh.GetId()] = monitorTcpMap
+				serviceTcpMap = make(map[uint64]*pingStore)
+				ss.serviceResponsePing[mh.GetId()] = serviceTcpMap
 			}
-			ts, ok := monitorTcpMap[r.Reporter]
+			ts, ok := serviceTcpMap[r.Reporter]
 			if !ok {
 				ts = &pingStore{}
 			}
 			ts.count++
 			ts.ping = (ts.ping*float32(ts.count-1) + mh.Delay) / float32(ts.count)
 			if ts.count == Conf.AvgPingCount {
-				if ts.ping > float32(Conf.MaxTCPPingValue) {
-					ts.ping = float32(Conf.MaxTCPPingValue)
-				}
 				ts.count = 0
-				if err := DB.Create(&model.MonitorHistory{
-					MonitorID: mh.GetId(),
+				if err := DB.Create(&model.ServiceHistory{
+					ServiceID: mh.GetId(),
 					AvgDelay:  ts.ping,
 					Data:      mh.Data,
 					ServerID:  r.Reporter,
 				}).Error; err != nil {
-					log.Println("NEZHA>> 服务监控数据持久化失败：", err)
+					log.Printf("NEZHA>> Failed to save service monitor metrics: %v", err)
 				}
 			}
-			monitorTcpMap[r.Reporter] = ts
+			serviceTcpMap[r.Reporter] = ts
 		}
 		ss.serviceResponseDataStoreLock.Lock()
 		// 写入当天状态
@@ -422,94 +494,35 @@ func (ss *ServiceSentinel) worker() {
 				index: 0,
 				t:     currentTime,
 			}
-			if err := DB.Create(&model.MonitorHistory{
-				MonitorID: mh.GetId(),
+			if err := DB.Create(&model.ServiceHistory{
+				ServiceID: mh.GetId(),
 				AvgDelay:  ss.serviceResponseDataStoreCurrentAvgDelay[mh.GetId()],
 				Data:      mh.Data,
 				Up:        ss.serviceResponseDataStoreCurrentUp[mh.GetId()],
 				Down:      ss.serviceResponseDataStoreCurrentDown[mh.GetId()],
 			}).Error; err != nil {
-				log.Println("NEZHA>> 服务监控数据持久化失败：", err)
+				log.Printf("NEZHA>> Failed to save service monitor metrics: %v", err)
 			}
 		}
 
+		cs, _ := ss.Get(mh.GetId())
+		m := ss.serverc.GetList()
 		// 延迟报警
 		if mh.Delay > 0 {
-			ss.monitorsLock.RLock()
-			if ss.monitors[mh.GetId()].LatencyNotify {
-				notificationTag := ss.monitors[mh.GetId()].NotificationTag
-				minMuteLabel := NotificationMuteLabel.ServiceLatencyMin(mh.GetId())
-				maxMuteLabel := NotificationMuteLabel.ServiceLatencyMax(mh.GetId())
-				if mh.Delay > ss.monitors[mh.GetId()].MaxLatency {
-					// 延迟超过最大值
-					ServerLock.RLock()
-					reporterServer := ServerList[r.Reporter]
-					msg := fmt.Sprintf("[Latency] %s %2f > %2f, Reporter: %s", ss.monitors[mh.GetId()].Name, mh.Delay, ss.monitors[mh.GetId()].MaxLatency, reporterServer.Name)
-					go SendNotification(notificationTag, msg, minMuteLabel)
-					ServerLock.RUnlock()
-				} else if mh.Delay < ss.monitors[mh.GetId()].MinLatency {
-					// 延迟低于最小值
-					ServerLock.RLock()
-					reporterServer := ServerList[r.Reporter]
-					msg := fmt.Sprintf("[Latency] %s %2f < %2f, Reporter: %s", ss.monitors[mh.GetId()].Name, mh.Delay, ss.monitors[mh.GetId()].MinLatency, reporterServer.Name)
-					go SendNotification(notificationTag, msg, maxMuteLabel)
-					ServerLock.RUnlock()
-				} else {
-					// 正常延迟， 清除静音缓存
-					UnMuteNotification(notificationTag, minMuteLabel)
-					UnMuteNotification(notificationTag, maxMuteLabel)
-				}
-			}
-			ss.monitorsLock.RUnlock()
+			delayCheck(&r, ss.notificationc, m, cs, mh)
 		}
 
 		// 状态变更报警+触发任务执行
 		if stateCode == StatusDown || stateCode != ss.lastStatus[mh.GetId()] {
-			ss.monitorsLock.Lock()
 			lastStatus := ss.lastStatus[mh.GetId()]
 			// 存储新的状态值
 			ss.lastStatus[mh.GetId()] = stateCode
 
-			// 判断是否需要发送通知
-			isNeedSendNotification := ss.monitors[mh.GetId()].Notify && (lastStatus != 0 || stateCode == StatusDown)
-			if isNeedSendNotification {
-				ServerLock.RLock()
-
-				reporterServer := ServerList[r.Reporter]
-				notificationTag := ss.monitors[mh.GetId()].NotificationTag
-				notificationMsg := fmt.Sprintf("[%s] %s Reporter: %s, Error: %s", StatusCodeToString(stateCode), ss.monitors[mh.GetId()].Name, reporterServer.Name, mh.Data)
-				muteLabel := NotificationMuteLabel.ServiceStateChanged(mh.GetId())
-
-				// 状态变更时，清除静音缓存
-				if stateCode != lastStatus {
-					UnMuteNotification(notificationTag, muteLabel)
-				}
-
-				go SendNotification(notificationTag, notificationMsg, muteLabel)
-				ServerLock.RUnlock()
-			}
-
-			// 判断是否需要触发任务
-			isNeedTriggerTask := ss.monitors[mh.GetId()].EnableTriggerTask && lastStatus != 0
-			if isNeedTriggerTask {
-				ServerLock.RLock()
-				reporterServer := ServerList[r.Reporter]
-				ServerLock.RUnlock()
-
-				if stateCode == StatusGood && lastStatus != stateCode {
-					// 当前状态正常 前序状态非正常时 触发恢复任务
-					go SendTriggerTasks(ss.monitors[mh.GetId()].RecoverTriggerTasks, reporterServer.ID)
-				} else if lastStatus == StatusGood && lastStatus != stateCode {
-					// 前序状态正常 当前状态非正常时 触发失败任务
-					go SendTriggerTasks(ss.monitors[mh.GetId()].FailTriggerTasks, reporterServer.ID)
-				}
-			}
-
-			ss.monitorsLock.Unlock()
+			notifyCheck(&r, ss.notificationc, ss.crc, m, cs, mh, lastStatus, stateCode)
 		}
 		ss.serviceResponseDataStoreLock.Unlock()
 
-		// SSL 证书报警
+		// TLS 证书报警
 		var errMsg string
 		if strings.HasPrefix(mh.Data, "SSL证书错误：") {
 			// i/o timeout、connection timeout、EOF 错误
@@ -517,29 +530,25 @@ func (ss *ServiceSentinel) worker() {
 				!strings.HasSuffix(mh.Data, "EOF") &&
 				!strings.HasSuffix(mh.Data, "timed out") {
 				errMsg = mh.Data
-				ss.monitorsLock.RLock()
-				if ss.monitors[mh.GetId()].Notify {
-					muteLabel := NotificationMuteLabel.ServiceSSL(mh.GetId(), "network")
-					go SendNotification(ss.monitors[mh.GetId()].NotificationTag, fmt.Sprintf("[SSL] Fetch cert info failed, %s %s", ss.monitors[mh.GetId()].Name, errMsg), muteLabel)
+				if cs.Notify {
+					muteLabel := NotificationMuteLabel.ServiceTLS(mh.GetId(), "network")
+					go ss.notificationc.SendNotification(cs.NotificationGroupID, Localizer.Tf("[TLS] Fetch cert info failed, Reporter: %s, Error: %s", cs.Name, errMsg), muteLabel)
 				}
-				ss.monitorsLock.RUnlock()
-
 			}
 		} else {
 			// 清除网络错误静音缓存
-			UnMuteNotification(ss.monitors[mh.GetId()].NotificationTag, NotificationMuteLabel.ServiceSSL(mh.GetId(), "network"))
+			ss.notificationc.UnMuteNotification(cs.NotificationGroupID, NotificationMuteLabel.ServiceTLS(mh.GetId(), "network"))
 
 			var newCert = strings.Split(mh.Data, "|")
 			if len(newCert) > 1 {
-				ss.monitorsLock.Lock()
-				enableNotify := ss.monitors[mh.GetId()].Notify
+				enableNotify := cs.Notify
 
 				// 首次获取证书信息时，缓存证书信息
-				if ss.sslCertCache[mh.GetId()] == "" {
-					ss.sslCertCache[mh.GetId()] = mh.Data
+				if ss.tlsCertCache[mh.GetId()] == "" {
+					ss.tlsCertCache[mh.GetId()] = mh.Data
 				}
 
-				oldCert := strings.Split(ss.sslCertCache[mh.GetId()], "|")
+				oldCert := strings.Split(ss.tlsCertCache[mh.GetId()], "|")
 				isCertChanged := false
 				expiresOld, _ := time.Parse("2006-01-02 15:04:05 -0700 MST", oldCert[1])
 				expiresNew, _ := time.Parse("2006-01-02 15:04:05 -0700 MST", newCert[1])
@@ -547,40 +556,96 @@ func (ss *ServiceSentinel) worker() {
 				// 证书变更时，更新缓存
 				if oldCert[0] != newCert[0] && !expiresNew.Equal(expiresOld) {
 					isCertChanged = true
-					ss.sslCertCache[mh.GetId()] = mh.Data
+					ss.tlsCertCache[mh.GetId()] = mh.Data
 				}
 
-				notificationTag := ss.monitors[mh.GetId()].NotificationTag
-				serviceName := ss.monitors[mh.GetId()].Name
-				ss.monitorsLock.Unlock()
+				notificationGroupID := cs.NotificationGroupID
+				serviceName := cs.Name
 
 				// 需要发送提醒
 				if enableNotify {
 					// 证书过期提醒
 					if expiresNew.Before(time.Now().AddDate(0, 0, 7)) {
 						expiresTimeStr := expiresNew.Format("2006-01-02 15:04:05")
-						errMsg = fmt.Sprintf(
-							"The SSL certificate will expire within seven days. Expiration time: %s",
+						errMsg = Localizer.Tf(
+							"The TLS certificate will expire within seven days. Expiration time: %s",
 							expiresTimeStr,
 						)
 
 						// 静音规则： 服务id+证书过期时间
 						// 用于避免多个监测点对相同证书同时报警
-						muteLabel := NotificationMuteLabel.ServiceSSL(mh.GetId(), fmt.Sprintf("expire_%s", expiresTimeStr))
-						go SendNotification(notificationTag, fmt.Sprintf("[SSL] %s %s", serviceName, errMsg), muteLabel)
+						muteLabel := NotificationMuteLabel.ServiceTLS(mh.GetId(), fmt.Sprintf("expire_%s", expiresTimeStr))
+						go ss.notificationc.SendNotification(notificationGroupID, fmt.Sprintf("[TLS] %s %s", serviceName, errMsg), muteLabel)
 					}
 
 					// 证书变更提醒
 					if isCertChanged {
-						errMsg = fmt.Sprintf(
-							"SSL certificate changed, old: %s, %s expired; new: %s, %s expired.",
+						errMsg = Localizer.Tf(
+							"TLS certificate changed, old: issuer %s, expires at %s; new: issuer %s, expires at %s",
 							oldCert[0], expiresOld.Format("2006-01-02 15:04:05"), newCert[0], expiresNew.Format("2006-01-02 15:04:05"))
 
 						// 证书变更后会自动更新缓存，所以不需要静音
-						go SendNotification(notificationTag, fmt.Sprintf("[SSL] %s %s", serviceName, errMsg), nil)
+						go ss.notificationc.SendNotification(notificationGroupID, fmt.Sprintf("[TLS] %s %s", serviceName, errMsg), nil)
 					}
 				}
 			}
+		}
+	}
+}
+
+func delayCheck(r *ReportData, nc *NotificationClass, m map[uint64]*model.Server, ss *model.Service, mh *pb.TaskResult) {
+	if !ss.LatencyNotify {
+		return
+	}
+
+	notificationGroupID := ss.NotificationGroupID
+	minMuteLabel := NotificationMuteLabel.ServiceLatencyMin(mh.GetId())
+	maxMuteLabel := NotificationMuteLabel.ServiceLatencyMax(mh.GetId())
+	if mh.Delay > ss.MaxLatency {
+		// 延迟超过最大值
+		reporterServer := m[r.Reporter]
+		msg := Localizer.Tf("[Latency] %s %2f > %2f, Reporter: %s", ss.Name, mh.Delay, ss.MaxLatency, reporterServer.Name)
+		go nc.SendNotification(notificationGroupID, msg, minMuteLabel)
+	} else if mh.Delay < ss.MinLatency {
+		// 延迟低于最小值
+		reporterServer := m[r.Reporter]
+		msg := Localizer.Tf("[Latency] %s %2f < %2f, Reporter: %s", ss.Name, mh.Delay, ss.MinLatency, reporterServer.Name)
+		go nc.SendNotification(notificationGroupID, msg, maxMuteLabel)
+	} else {
+		// 正常延迟， 清除静音缓存
+		nc.UnMuteNotification(notificationGroupID, minMuteLabel)
+		nc.UnMuteNotification(notificationGroupID, maxMuteLabel)
+	}
+}
+
+func notifyCheck(r *ReportData, nc *NotificationClass, crc *CronClass, m map[uint64]*model.Server,
+	ss *model.Service, mh *pb.TaskResult, lastStatus, stateCode uint8) {
+	// 判断是否需要发送通知
+	isNeedSendNotification := ss.Notify && (lastStatus != 0 || stateCode == StatusDown)
+	if isNeedSendNotification {
+		reporterServer := m[r.Reporter]
+		notificationGroupID := ss.NotificationGroupID
+		notificationMsg := Localizer.Tf("[%s] %s Reporter: %s, Error: %s", StatusCodeToString(stateCode), ss.Name, reporterServer.Name, mh.Data)
+		muteLabel := NotificationMuteLabel.ServiceStateChanged(mh.GetId())
+
+		// 状态变更时，清除静音缓存
+		if stateCode != lastStatus {
+			nc.UnMuteNotification(notificationGroupID, muteLabel)
+		}
+
+		go nc.SendNotification(notificationGroupID, notificationMsg, muteLabel)
+	}
+
+	// 判断是否需要触发任务
+	isNeedTriggerTask := ss.EnableTriggerTask && lastStatus != 0
+	if isNeedTriggerTask {
+		reporterServer := m[r.Reporter]
+		if stateCode == StatusGood && lastStatus != stateCode {
+			// 当前状态正常 前序状态非正常时 触发恢复任务
+			go crc.SendTriggerTasks(ss.RecoverTriggerTasks, reporterServer.ID)
+		} else if lastStatus == StatusGood && lastStatus != stateCode {
+			// 前序状态正常 当前状态非正常时 触发失败任务
+			go crc.SendTriggerTasks(ss.FailTriggerTasks, reporterServer.ID)
 		}
 	}
 }
@@ -593,7 +658,7 @@ const (
 	StatusDown
 )
 
-func GetStatusCode[T float32 | uint64](percent T) int {
+func GetStatusCode[T constraints.Float | constraints.Integer](percent T) uint8 {
 	if percent == 0 {
 		return StatusNoData
 	}
@@ -606,16 +671,16 @@ func GetStatusCode[T float32 | uint64](percent T) int {
 	return StatusDown
 }
 
-func StatusCodeToString(statusCode int) string {
+func StatusCodeToString(statusCode uint8) string {
 	switch statusCode {
 	case StatusNoData:
-		return Localizer.MustLocalize(&i18n.LocalizeConfig{MessageID: "StatusNoData"})
+		return Localizer.T("No Data")
 	case StatusGood:
-		return Localizer.MustLocalize(&i18n.LocalizeConfig{MessageID: "StatusGood"})
+		return Localizer.T("Good")
 	case StatusLowAvailability:
-		return Localizer.MustLocalize(&i18n.LocalizeConfig{MessageID: "StatusLowAvailability"})
+		return Localizer.T("Low Availability")
 	case StatusDown:
-		return Localizer.MustLocalize(&i18n.LocalizeConfig{MessageID: "StatusDown"})
+		return Localizer.T("Down")
 	default:
 		return ""
 	}
