@@ -2,100 +2,170 @@ package main
 
 import (
 	"context"
+	"embed"
+	"flag"
 	"fmt"
 	"log"
+	"net"
+	"net/http"
+	"os"
+	"strings"
 	"time"
 	_ "time/tzdata"
 
-	"github.com/naiba/nezha/cmd/dashboard/controller"
-	"github.com/naiba/nezha/cmd/dashboard/rpc"
-	"github.com/naiba/nezha/model"
-	"github.com/naiba/nezha/proto"
-	"github.com/naiba/nezha/service/singleton"
+	"github.com/gin-gonic/gin"
 	"github.com/ory/graceful"
-	flag "github.com/spf13/pflag"
+	"golang.org/x/crypto/bcrypt"
+	"golang.org/x/net/http2"
+	"golang.org/x/net/http2/h2c"
+
+	"github.com/nezhahq/nezha/cmd/dashboard/controller"
+	"github.com/nezhahq/nezha/cmd/dashboard/controller/waf"
+	"github.com/nezhahq/nezha/cmd/dashboard/rpc"
+	"github.com/nezhahq/nezha/model"
+	"github.com/nezhahq/nezha/proto"
+	"github.com/nezhahq/nezha/service/singleton"
 )
 
 type DashboardCliParam struct {
 	Version          bool   // 当前版本号
 	ConfigFile       string // 配置文件路径
-	DatebaseLocation string // Sqlite3 数据库文件路径
+	DatabaseLocation string // Sqlite3 数据库文件路径
 }
 
 var (
 	dashboardCliParam DashboardCliParam
+	//go:embed *-dist
+	frontendDist embed.FS
 )
 
-func init() {
-	flag.CommandLine.ParseErrorsWhitelist.UnknownFlags = true
-	flag.BoolVarP(&dashboardCliParam.Version, "version", "v", false, "查看当前版本号")
-	flag.StringVarP(&dashboardCliParam.ConfigFile, "config", "c", "data/config.yaml", "配置文件路径")
-	flag.StringVar(&dashboardCliParam.DatebaseLocation, "db", "data/sqlite.db", "Sqlite3数据库文件路径")
-	flag.Parse()
-
-	// 初始化 dao 包
-	singleton.InitConfigFromPath(dashboardCliParam.ConfigFile)
-	singleton.InitTimezoneAndCache()
-	singleton.InitDBFromPath(dashboardCliParam.DatebaseLocation)
-	singleton.InitLocalizer()
-	initSystem()
-}
-
 func initSystem() {
+	// 初始化管理员账户
+	var usersCount int64
+	if err := singleton.DB.Model(&model.User{}).Count(&usersCount).Error; err != nil {
+		panic(err)
+	}
+	if usersCount == 0 {
+		hash, err := bcrypt.GenerateFromPassword([]byte("admin"), bcrypt.DefaultCost)
+		if err != nil {
+			panic(err)
+		}
+		admin := model.User{
+			Username: "admin",
+			Password: string(hash),
+		}
+		if err := singleton.DB.Create(&admin).Error; err != nil {
+			panic(err)
+		}
+	}
+
 	// 启动 singleton 包下的所有服务
 	singleton.LoadSingleton()
 
 	// 每天的3:30 对 监控记录 和 流量记录 进行清理
-	if _, err := singleton.Cron.AddFunc("0 30 3 * * *", singleton.CleanMonitorHistory); err != nil {
+	if _, err := singleton.CronShared.AddFunc("0 30 3 * * *", singleton.CleanServiceHistory); err != nil {
 		panic(err)
 	}
 
 	// 每小时对流量记录进行打点
-	if _, err := singleton.Cron.AddFunc("0 0 * * * *", singleton.RecordTransferHourlyUsage); err != nil {
+	if _, err := singleton.CronShared.AddFunc("0 0 * * * *", singleton.RecordTransferHourlyUsage); err != nil {
 		panic(err)
 	}
 }
 
+// @title           Nezha Monitoring API
+// @version         1.0
+// @description     Nezha Monitoring API
+// @termsOfService  http://nezhahq.github.io
+
+// @contact.name   API Support
+// @contact.url    http://nezhahq.github.io
+// @contact.email  hi@nai.ba
+
+// @license.name  Apache 2.0
+// @license.url   http://www.apache.org/licenses/LICENSE-2.0.html
+
+// @host      localhost:8008
+// @BasePath  /api/v1
+
+// @securityDefinitions.apikey  BearerAuth
+// @in header
+// @name Authorization
+
+// @externalDocs.description  OpenAPI
+// @externalDocs.url          https://swagger.io/resources/open-api/
 func main() {
+	flag.BoolVar(&dashboardCliParam.Version, "v", false, "查看当前版本号")
+	flag.StringVar(&dashboardCliParam.ConfigFile, "c", "data/config.yaml", "配置文件路径")
+	flag.StringVar(&dashboardCliParam.DatabaseLocation, "db", "data/sqlite.db", "Sqlite3数据库文件路径")
+	flag.Parse()
+
 	if dashboardCliParam.Version {
 		fmt.Println(singleton.Version)
-		return
+		os.Exit(0)
 	}
 
-	// TODO 使用 cmux 在同一端口服务 HTTP 和 gRPC
-	singleton.CleanMonitorHistory()
-	go rpc.ServeRPC(singleton.Conf.GRPCPort)
-	serviceSentinelDispatchBus := make(chan model.Monitor) // 用于传递服务监控任务信息的channel
+	// 初始化 dao 包
+	singleton.InitFrontendTemplates()
+	singleton.InitConfigFromPath(dashboardCliParam.ConfigFile)
+	singleton.InitTimezoneAndCache()
+	singleton.InitDBFromPath(dashboardCliParam.DatabaseLocation)
+	initSystem()
+
+	l, err := net.Listen("tcp", fmt.Sprintf("%s:%d", singleton.Conf.ListenHost, singleton.Conf.ListenPort))
+	if err != nil {
+		log.Fatal(err)
+	}
+
+	singleton.CleanServiceHistory()
+	serviceSentinelDispatchBus := make(chan *model.Service) // 用于传递服务监控任务信息的channel
+	rpc.DispatchKeepalive()
 	go rpc.DispatchTask(serviceSentinelDispatchBus)
-	go rpc.DispatchKeepalive()
 	go singleton.AlertSentinelStart()
-	singleton.NewServiceSentinel(serviceSentinelDispatchBus)
-	srv := controller.ServeWeb(singleton.Conf.HTTPPort)
-	go dispatchReportInfoTask()
+	singleton.ServiceSentinelShared, err = singleton.NewServiceSentinel(
+		serviceSentinelDispatchBus, singleton.ServerShared, singleton.NotificationShared, singleton.CronShared)
+	if err != nil {
+		log.Fatal(err)
+	}
+
+	grpcHandler := rpc.ServeRPC()
+	httpHandler := controller.ServeWeb(frontendDist)
+	controller.InitUpgrader()
+
+	muxHandler := newHTTPandGRPCMux(httpHandler, grpcHandler)
+	http2Server := &http2.Server{}
+	muxServer := &http.Server{Handler: h2c.NewHandler(muxHandler, http2Server), ReadHeaderTimeout: time.Second * 5}
+
 	if err := graceful.Graceful(func() error {
-		return srv.ListenAndServe()
+		log.Printf("NEZHA>> Dashboard::START ON %s:%d", singleton.Conf.ListenHost, singleton.Conf.ListenPort)
+		return muxServer.Serve(l)
 	}, func(c context.Context) error {
 		log.Println("NEZHA>> Graceful::START")
 		singleton.RecordTransferHourlyUsage()
 		log.Println("NEZHA>> Graceful::END")
-		srv.Shutdown(c)
-		return nil
+		return muxServer.Shutdown(c)
 	}); err != nil {
 		log.Printf("NEZHA>> ERROR: %v", err)
 	}
 }
 
-func dispatchReportInfoTask() {
-	time.Sleep(time.Second * 15)
-	singleton.ServerLock.RLock()
-	defer singleton.ServerLock.RUnlock()
-	for _, server := range singleton.ServerList {
-		if server == nil || server.TaskStream == nil {
-			continue
+func newHTTPandGRPCMux(httpHandler http.Handler, grpcHandler http.Handler) http.Handler {
+	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+		natConfig := singleton.NATShared.GetNATConfigByDomain(r.Host)
+		if natConfig != nil {
+			if !natConfig.Enabled {
+				c, _ := gin.CreateTestContext(w)
+				waf.ShowBlockPage(c, fmt.Errorf("nat host %s is disabled", natConfig.Domain))
+				return
+			}
+			rpc.ServeNAT(w, r, natConfig)
+			return
 		}
-		server.TaskStream.Send(&proto.Task{
-			Type: model.TaskTypeReportHostInfo,
-			Data: "",
-		})
-	}
+		if r.ProtoMajor == 2 && r.Header.Get("Content-Type") == "application/grpc" &&
+			strings.HasPrefix(r.URL.Path, "/"+proto.NezhaService_ServiceDesc.ServiceName) {
+			grpcHandler.ServeHTTP(w, r)
+			return
+		}
+		httpHandler.ServeHTTP(w, r)
+	})
 }
