@@ -1,9 +1,7 @@
 package model
 
 import (
-	"time"
-
-	"github.com/naiba/nezha/pkg/utils"
+	"github.com/nezhahq/nezha/pkg/utils"
 	"gorm.io/gorm"
 )
 
@@ -12,29 +10,18 @@ const (
 	ModeOnetimeTrigger = 1
 )
 
-type CycleTransferStats struct {
-	Name       string
-	From       time.Time
-	To         time.Time
-	Max        uint64
-	Min        uint64
-	ServerName map[uint64]string
-	Transfer   map[uint64]uint64
-	NextUpdate map[uint64]time.Time
-}
-
 type AlertRule struct {
 	Common
-	Name                   string
-	RulesRaw               string
-	Enable                 *bool
-	TriggerMode            int      `gorm:"default:0"` // 触发模式: 0-始终触发(默认) 1-单次触发
-	NotificationTag        string   // 该报警规则所在的通知组
-	FailTriggerTasksRaw    string   `gorm:"default:'[]'"`
-	RecoverTriggerTasksRaw string   `gorm:"default:'[]'"`
-	Rules                  []Rule   `gorm:"-" json:"-"`
-	FailTriggerTasks       []uint64 `gorm:"-" json:"-"` // 失败时执行的触发任务id
-	RecoverTriggerTasks    []uint64 `gorm:"-" json:"-"` // 恢复时执行的触发任务id
+	Name                   string   `json:"name"`
+	RulesRaw               string   `json:"-"`
+	Enable                 *bool    `json:"enable,omitempty"`
+	TriggerMode            uint8    `gorm:"default:0" json:"trigger_mode"` // 触发模式: 0-始终触发(默认) 1-单次触发
+	NotificationGroupID    uint64   `json:"notification_group_id"`         // 该报警规则所在的通知组
+	FailTriggerTasksRaw    string   `gorm:"default:'[]'" json:"-"`
+	RecoverTriggerTasksRaw string   `gorm:"default:'[]'" json:"-"`
+	Rules                  []*Rule  `gorm:"-" json:"rules"`
+	FailTriggerTasks       []uint64 `gorm:"-" json:"fail_trigger_tasks"`    // 失败时执行的触发任务id
+	RecoverTriggerTasks    []uint64 `gorm:"-" json:"recover_trigger_tasks"` // 恢复时执行的触发任务id
 }
 
 func (r *AlertRule) BeforeSave(tx *gorm.DB) error {
@@ -74,55 +61,61 @@ func (r *AlertRule) Enabled() bool {
 	return r.Enable != nil && *r.Enable
 }
 
-// Snapshot 对传入的Server进行该报警规则下所有type的检查 返回包含每项检查结果的空接口
-func (r *AlertRule) Snapshot(cycleTransferStats *CycleTransferStats, server *Server, db *gorm.DB) []interface{} {
-	var point []interface{}
-	for i := 0; i < len(r.Rules); i++ {
-		point = append(point, r.Rules[i].Snapshot(cycleTransferStats, server, db))
+// Snapshot 对传入的Server进行该报警规则下所有type的检查 返回每项检查结果
+func (r *AlertRule) Snapshot(cycleTransferStats *CycleTransferStats, server *Server, db *gorm.DB) []bool {
+	point := make([]bool, len(r.Rules))
+
+	for i, rule := range r.Rules {
+		point[i] = rule.Snapshot(cycleTransferStats, server, db)
 	}
 	return point
 }
 
-// Check 传入包含当前报警规则下所有type检查结果的空接口 返回报警持续时间与是否通过报警检查(通过则返回true)
-func (r *AlertRule) Check(points [][]interface{}) (int, bool) {
-	var maxNum int // 报警持续时间
-	var count int  // 检查未通过的个数
-	for i := 0; i < len(r.Rules); i++ {
-		if r.Rules[i].IsTransferDurationRule() {
+// Check 传入包含当前报警规则下所有type检查结果 返回报警持续时间与是否通过报警检查(通过则返回true)
+func (r *AlertRule) Check(points [][]bool) (maxDuration int, passed bool) {
+	var hasPassedRule bool
+
+	for ruleId, rule := range r.Rules {
+		if rule.IsTransferDurationRule() {
 			// 循环区间流量报警
-			if maxNum < 1 {
-				maxNum = 1
+			if maxDuration < 1 {
+				maxDuration = 1
 			}
-			for j := len(points[i]) - 1; j >= 0; j-- {
-				if points[i][j] != nil {
-					count++
-					break
-				}
+			if hasPassedRule {
+				continue
+			}
+			// 只要最后一次检查超出了规则范围 就认为检查未通过
+			if len(points) > 0 && points[len(points)-1][ruleId] {
+				hasPassedRule = true
 			}
 		} else {
 			// 常规报警
-			total := 0.0
-			fail := 0.0
-			num := int(r.Rules[i].Duration)
-			if num > maxNum {
-				maxNum = num
+			duration := int(rule.Duration)
+			if duration > maxDuration {
+				maxDuration = duration
 			}
-			if len(points) < num {
+			if hasPassedRule {
 				continue
 			}
-			for j := len(points) - 1; j >= 0 && len(points)-num <= j; j-- {
+			if len(points) < duration {
+				// 如果采样点数量不足 则认为检查通过
+				hasPassedRule = true
+				continue
+			}
+			total, fail := 0, 0
+			for timeTick := len(points) - duration; timeTick < len(points); timeTick++ {
 				total++
-				if points[j][i] != nil {
+				if !points[timeTick][ruleId] {
 					fail++
 				}
 			}
 			// 当70%以上的采样点未通过规则判断时 才认为当前检查未通过
-			if fail/total > 0.7 {
-				count++
-				break
+			if fail*100/total <= 70 {
+				hasPassedRule = true
 			}
 		}
 	}
-	// 仅当所有检查均未通过时 返回false
-	return maxNum, count != len(r.Rules)
+
+	// 仅当所有检查均未通过时 才触发告警
+	return maxDuration, hasPassedRule
 }
