@@ -1,185 +1,157 @@
 package model
 
 import (
+	"maps"
 	"os"
+	"path/filepath"
+	"slices"
 	"strconv"
 	"strings"
 
-	"github.com/spf13/viper"
-	"sigs.k8s.io/yaml"
-)
+	kyaml "github.com/knadh/koanf/parsers/yaml"
+	"github.com/knadh/koanf/providers/env"
+	"github.com/knadh/koanf/providers/file"
+	"github.com/knadh/koanf/v2"
+	"gopkg.in/yaml.v3"
 
-var Languages = map[string]string{
-	"zh-CN": "简体中文",
-	"zh-TW": "繁體中文",
-	"en-US": "English",
-	"es-ES": "Español",
-}
-
-var Themes = map[string]string{
-	"default":       "Default",
-	"daynight":      "JackieSung DayNight",
-	"mdui":          "Neko Mdui",
-	"hotaru":        "Hotaru",
-	"angel-kanade":  "AngelKanade",
-	"server-status": "ServerStatus",
-	"custom":        "Custom(local)",
-}
-
-var DashboardThemes = map[string]string{
-	"default": "Default",
-	"custom":  "Custom(local)",
-}
-
-const (
-	ConfigTypeGitHub     = "github"
-	ConfigTypeGitee      = "gitee"
-	ConfigTypeGitlab     = "gitlab"
-	ConfigTypeJihulab    = "jihulab"
-	ConfigTypeGitea      = "gitea"
-	ConfigTypeCloudflare = "cloudflare"
-	ConfigTypeOidc       = "oidc"
+	"github.com/nezhahq/nezha/pkg/utils"
 )
 
 const (
-	ConfigCoverAll = iota
+	ConfigUsePeerIP = "NZ::Use-Peer-IP"
+	ConfigCoverAll  = iota
 	ConfigCoverIgnoreAll
 )
 
-type AgentConfig struct {
-	HardDrivePartitionAllowlist []string
-	NICAllowlist                map[string]bool
-	v                           *viper.Viper
+type ConfigForGuests struct {
+	Language            string   `json:"language"`
+	SiteName            string   `json:"site_name"`
+	CustomCode          string   `json:"custom_code,omitempty"`
+	CustomCodeDashboard string   `json:"custom_code_dashboard,omitempty"`
+	Oauth2Providers     []string `json:"oauth2_providers,omitempty"`
+
+	InstallHost string `json:"install_host,omitempty"`
+	TLS         bool   `json:"tls,omitempty"`
 }
 
-// Read 从给定的文件目录加载配置文件
-func (c *AgentConfig) Read(path string) error {
-	c.v = viper.New()
-	c.v.SetConfigFile(path)
-	err := c.v.ReadInConfig()
-	if err != nil {
-		return err
-	}
-	err = c.v.Unmarshal(c)
-	if err != nil {
-		return err
-	}
-	return nil
-}
-
-func (c *AgentConfig) Save() error {
-	data, err := yaml.Marshal(c)
-	if err != nil {
-		return err
-	}
-	return os.WriteFile(c.v.ConfigFileUsed(), data, 0600)
-}
-
-// Config 站点配置
 type Config struct {
-	Debug    bool   // debug模式开关
-	Language string // 系统语言，默认 zh-CN
-	Site     struct {
-		Brand               string // 站点名称
-		CookieName          string // 浏览器 Cookie 名称
-		Theme               string
-		DashboardTheme      string
-		CustomCode          string
-		CustomCodeDashboard string
-		ViewPassword        string // 前台查看密码
-	}
-	Oauth2 struct {
-		Type            string
-		Admin           string // 管理员用户名列表
-		AdminGroups     string // 管理员用户组列表
-		ClientID        string
-		ClientSecret    string
-		Endpoint        string
-		OidcDisplayName string // for OIDC Display Name
-		OidcIssuer      string // for OIDC Issuer
-		OidcLogoutURL   string // for OIDC Logout URL
-		OidcRegisterURL string // for OIDC Register URL
-		OidcLoginClaim  string // for OIDC Claim
-		OidcGroupClaim  string // for OIDC Group Claim
-		OidcScopes      string // for OIDC Scopes
-		OidcAutoCreate  bool   // for OIDC Auto Create
-		OidcAutoLogin   bool   // for OIDC Auto Login
-	}
-	HTTPPort      uint
-	GRPCPort      uint
-	GRPCHost      string
-	ProxyGRPCPort uint
-	TLS           bool
+	Debug        bool   `mapstructure:"debug" json:"debug,omitempty"`                   // debug模式开关
+	RealIPHeader string `mapstructure:"real_ip_header" json:"real_ip_header,omitempty"` // 真实IP
 
-	EnablePlainIPInNotification     bool // 通知信息IP不打码
-	DisableSwitchTemplateInFrontend bool // 前台禁用切换模板功能
+	Language       string `mapstructure:"language" json:"language"` // 系统语言，默认 zh_CN
+	SiteName       string `mapstructure:"site_name" json:"site_name"`
+	UserTemplate   string `mapstructure:"user_template" json:"user_template,omitempty"`
+	AdminTemplate  string `mapstructure:"admin_template" json:"admin_template,omitempty"`
+	JWTSecretKey   string `mapstructure:"jwt_secret_key" json:"jwt_secret_key,omitempty"`
+	AgentSecretKey string `mapstructure:"agent_secret_key" json:"agent_secret_key,omitempty"`
+	ListenPort     uint   `mapstructure:"listen_port" json:"listen_port,omitempty"`
+	ListenHost     string `mapstructure:"listen_host" json:"listen_host,omitempty"`
+	InstallHost    string `mapstructure:"install_host" json:"install_host,omitempty"`
+	TLS            bool   `mapstructure:"tls" json:"tls,omitempty"`
+	Location       string `mapstructure:"location" json:"location,omitempty"`     // 时区，默认为 Asia/Shanghai
+	ForceAuth      bool   `mapstructure:"force_auth" json:"force_auth,omitempty"` // 强制要求认证
+
+	EnablePlainIPInNotification bool `mapstructure:"enable_plain_ip_in_notification" json:"enable_plain_ip_in_notification,omitempty"` // 通知信息IP不打码
 
 	// IP变更提醒
-	EnableIPChangeNotification bool
-	IPChangeNotificationTag    string
-	Cover                      uint8  // 覆盖范围（0:提醒未被 IgnoredIPNotification 包含的所有服务器; 1:仅提醒被 IgnoredIPNotification 包含的服务器;）
-	IgnoredIPNotification      string // 特定服务器IP（多个服务器用逗号分隔）
+	EnableIPChangeNotification  bool   `mapstructure:"enable_ip_change_notification" json:"enable_ip_change_notification,omitempty"`
+	IPChangeNotificationGroupID uint64 `mapstructure:"ip_change_notification_group_id" json:"ip_change_notification_group_id"`
+	Cover                       uint8  `mapstructure:"cover" json:"cover"`                                               // 覆盖范围（0:提醒未被 IgnoredIPNotification 包含的所有服务器; 1:仅提醒被 IgnoredIPNotification 包含的服务器;）
+	IgnoredIPNotification       string `mapstructure:"ignored_ip_notification" json:"ignored_ip_notification,omitempty"` // 特定服务器IP（多个服务器用逗号分隔）
 
-	Location string // 时区，默认为 Asia/Shanghai
+	IgnoredIPNotificationServerIDs map[uint64]bool `mapstructure:"ignored_ip_notification_server_ids" json:"ignored_ip_notification_server_ids,omitempty"` // [ServerID] -> bool(值为true代表当前ServerID在特定服务器列表内）
+	AvgPingCount                   int             `mapstructure:"avg_ping_count" json:"avg_ping_count,omitempty"`
+	DNSServers                     string          `mapstructure:"dns_servers" json:"dns_servers,omitempty"`
 
-	v                              *viper.Viper
-	IgnoredIPNotificationServerIDs map[uint64]bool // [ServerID] -> bool(值为true代表当前ServerID在特定服务器列表内）
-	MaxTCPPingValue                int32
-	AvgPingCount                   int
+	CustomCode          string `mapstructure:"custom_code" json:"custom_code,omitempty"`
+	CustomCodeDashboard string `mapstructure:"custom_code_dashboard" json:"custom_code_dashboard,omitempty"`
 
-	DNSServers string
+	// oauth2 配置
+	Oauth2 map[string]*Oauth2Config `mapstructure:"oauth2" json:"oauth2,omitempty"`
+	// oauth2 供应商列表，无需配置，自动生成
+	Oauth2Providers []string `yaml:"-" json:"oauth2_providers,omitempty"`
+
+	k        *koanf.Koanf `json:"-"`
+	filePath string       `json:"-"`
 }
 
 // Read 读取配置文件并应用
-func (c *Config) Read(path string) error {
-	c.v = viper.New()
-	c.v.SetConfigFile(path)
-	err := c.v.ReadInConfig()
+func (c *Config) Read(path string, frontendTemplates []FrontendTemplate) error {
+	c.k = koanf.New(".")
+	c.filePath = path
+
+	err := c.k.Load(env.Provider("NZ_", ".", func(s string) string {
+		return strings.Replace(strings.ToLower(strings.TrimPrefix(s, "NZ_")), "_", ".", -1)
+	}), nil)
 	if err != nil {
 		return err
 	}
 
-	err = c.v.Unmarshal(c)
+	if _, err := os.Stat(path); err == nil {
+		err = c.k.Load(file.Provider(path), kyaml.Parser())
+		if err != nil {
+			return err
+		}
+	}
+
+	err = c.k.Unmarshal("", c)
 	if err != nil {
 		return err
 	}
-
-	if c.Site.Theme == "" {
-		c.Site.Theme = "default"
-	}
-	if c.Site.DashboardTheme == "" {
-		c.Site.DashboardTheme = "default"
+	if c.ListenPort == 0 {
+		c.ListenPort = 8008
 	}
 	if c.Language == "" {
-		c.Language = "zh-CN"
-	}
-	if c.GRPCPort == 0 {
-		c.GRPCPort = 5555
-	}
-	if c.EnableIPChangeNotification && c.IPChangeNotificationTag == "" {
-		c.IPChangeNotificationTag = "default"
+		c.Language = "en_US"
 	}
 	if c.Location == "" {
 		c.Location = "Asia/Shanghai"
 	}
-	if c.MaxTCPPingValue == 0 {
-		c.MaxTCPPingValue = 1000
+	var userTemplateValid, adminTemplateValid bool
+	for _, v := range frontendTemplates {
+		if !userTemplateValid && v.Path == c.UserTemplate && !v.IsAdmin {
+			userTemplateValid = true
+		}
+		if !adminTemplateValid && v.Path == c.AdminTemplate && v.IsAdmin {
+			adminTemplateValid = true
+		}
+		if userTemplateValid && adminTemplateValid {
+			break
+		}
+	}
+	if c.UserTemplate == "" || !userTemplateValid {
+		c.UserTemplate = "user-dist"
+	}
+	if c.AdminTemplate == "" || !adminTemplateValid {
+		c.AdminTemplate = "admin-dist"
 	}
 	if c.AvgPingCount == 0 {
 		c.AvgPingCount = 2
 	}
-	if c.Oauth2.OidcScopes == "" {
-		c.Oauth2.OidcScopes = "openid,profile,email"
+	if c.Cover == 0 {
+		c.Cover = 1
 	}
-	if c.Oauth2.OidcLoginClaim == "" {
-		c.Oauth2.OidcLoginClaim = "sub"
+	if c.JWTSecretKey == "" {
+		c.JWTSecretKey, err = utils.GenerateRandomString(1024)
+		if err != nil {
+			return err
+		}
+		if err = c.Save(); err != nil {
+			return err
+		}
 	}
-	if c.Oauth2.OidcDisplayName == "" {
-		c.Oauth2.OidcDisplayName = "OIDC"
+
+	if c.AgentSecretKey == "" {
+		c.AgentSecretKey, err = utils.GenerateRandomString(32)
+		if err != nil {
+			return err
+		}
+		if err = c.Save(); err != nil {
+			return err
+		}
 	}
-	if c.Oauth2.OidcGroupClaim == "" {
-		c.Oauth2.OidcGroupClaim = "groups"
-	}
+
+	c.Oauth2Providers = slices.Collect(maps.Keys(c.Oauth2))
 
 	c.updateIgnoredIPNotificationID()
 	return nil
@@ -204,5 +176,11 @@ func (c *Config) Save() error {
 	if err != nil {
 		return err
 	}
-	return os.WriteFile(c.v.ConfigFileUsed(), data, 0600)
+
+	dir := filepath.Dir(c.filePath)
+	if err := os.MkdirAll(dir, 0750); err != nil {
+		return err
+	}
+
+	return os.WriteFile(c.filePath, data, 0600)
 }
