@@ -1,7 +1,9 @@
 package model
 
 import (
-	pb "github.com/naiba/nezha/proto"
+	"fmt"
+
+	pb "github.com/nezhahq/nezha/proto"
 )
 
 const (
@@ -16,23 +18,23 @@ type SensorTemperature struct {
 }
 
 type HostState struct {
-	CPU            float64
-	MemUsed        uint64
-	SwapUsed       uint64
-	DiskUsed       uint64
-	NetInTransfer  uint64
-	NetOutTransfer uint64
-	NetInSpeed     uint64
-	NetOutSpeed    uint64
-	Uptime         uint64
-	Load1          float64
-	Load5          float64
-	Load15         float64
-	TcpConnCount   uint64
-	UdpConnCount   uint64
-	ProcessCount   uint64
-	Temperatures   []SensorTemperature
-	GPU            float64
+	CPU            float64             `json:"cpu,omitempty"`
+	MemUsed        uint64              `json:"mem_used,omitempty"`
+	SwapUsed       uint64              `json:"swap_used,omitempty"`
+	DiskUsed       uint64              `json:"disk_used,omitempty"`
+	NetInTransfer  uint64              `json:"net_in_transfer,omitempty"`
+	NetOutTransfer uint64              `json:"net_out_transfer,omitempty"`
+	NetInSpeed     uint64              `json:"net_in_speed,omitempty"`
+	NetOutSpeed    uint64              `json:"net_out_speed,omitempty"`
+	Uptime         uint64              `json:"uptime,omitempty"`
+	Load1          float64             `json:"load_1,omitempty"`
+	Load5          float64             `json:"load_5,omitempty"`
+	Load15         float64             `json:"load_15,omitempty"`
+	TcpConnCount   uint64              `json:"tcp_conn_count,omitempty"`
+	UdpConnCount   uint64              `json:"udp_conn_count,omitempty"`
+	ProcessCount   uint64              `json:"process_count,omitempty"`
+	Temperatures   []SensorTemperature `json:"temperatures,omitempty"`
+	GPU            []float64           `json:"gpu,omitempty"`
 }
 
 func (s *HostState) PB() *pb.State {
@@ -96,19 +98,17 @@ func PB2State(s *pb.State) HostState {
 }
 
 type Host struct {
-	Platform        string
-	PlatformVersion string
-	CPU             []string
-	MemTotal        uint64
-	DiskTotal       uint64
-	SwapTotal       uint64
-	Arch            string
-	Virtualization  string
-	BootTime        uint64
-	IP              string `json:"-"`
-	CountryCode     string
-	Version         string
-	GPU             []string
+	Platform        string   `json:"platform,omitempty"`
+	PlatformVersion string   `json:"platform_version,omitempty"`
+	CPU             []string `json:"cpu,omitempty"`
+	MemTotal        uint64   `json:"mem_total,omitempty"`
+	DiskTotal       uint64   `json:"disk_total,omitempty"`
+	SwapTotal       uint64   `json:"swap_total,omitempty"`
+	Arch            string   `json:"arch,omitempty"`
+	Virtualization  string   `json:"virtualization,omitempty"`
+	BootTime        uint64   `json:"boot_time,omitempty"`
+	Version         string   `json:"version,omitempty"`
+	GPU             []string `json:"gpu,omitempty"`
 }
 
 func (h *Host) PB() *pb.Host {
@@ -122,10 +122,23 @@ func (h *Host) PB() *pb.Host {
 		Arch:            h.Arch,
 		Virtualization:  h.Virtualization,
 		BootTime:        h.BootTime,
-		Ip:              h.IP,
-		CountryCode:     h.CountryCode,
 		Version:         h.Version,
 		Gpu:             h.GPU,
+	}
+}
+
+// Filter returns a new instance of Host with some fields redacted.
+func (h *Host) Filter() *Host {
+	return &Host{
+		Platform:       h.Platform,
+		CPU:            h.CPU,
+		MemTotal:       h.MemTotal,
+		DiskTotal:      h.DiskTotal,
+		SwapTotal:      h.SwapTotal,
+		Arch:           h.Arch,
+		Virtualization: h.Virtualization,
+		BootTime:       h.BootTime,
+		GPU:            h.GPU,
 	}
 }
 
@@ -140,9 +153,36 @@ func PB2Host(h *pb.Host) Host {
 		Arch:            h.GetArch(),
 		Virtualization:  h.GetVirtualization(),
 		BootTime:        h.GetBootTime(),
-		IP:              h.GetIp(),
-		CountryCode:     h.GetCountryCode(),
 		Version:         h.GetVersion(),
 		GPU:             h.GetGpu(),
+	}
+}
+
+type IP struct {
+	IPv4Addr string `json:"ipv4_addr,omitempty"`
+	IPv6Addr string `json:"ipv6_addr,omitempty"`
+}
+
+func (p *IP) Join() string {
+	if p.IPv4Addr != "" && p.IPv6Addr != "" {
+		return fmt.Sprintf("%s/%s", p.IPv4Addr, p.IPv6Addr)
+	} else if p.IPv4Addr != "" {
+		return p.IPv4Addr
+	}
+	return p.IPv6Addr
+}
+
+type GeoIP struct {
+	IP          IP     `json:"ip,omitempty"`
+	CountryCode string `json:"country_code,omitempty"`
+}
+
+func PB2GeoIP(p *pb.GeoIP) GeoIP {
+	pbIP := p.GetIp()
+	return GeoIP{
+		IP: IP{
+			IPv4Addr: pbIP.GetIpv4(),
+			IPv6Addr: pbIP.GetIpv6(),
+		},
 	}
 }
