@@ -2,8 +2,9 @@ package model
 
 type NAT struct {
 	Common
-	Name     string
-	ServerID uint64
-	Host     string
-	Domain   string `gorm:"unique"`
+	Enabled  bool   `json:"enabled"`
+	Name     string `json:"name"`
+	ServerID uint64 `json:"server_id"`
+	Host     string `json:"host"`
+	Domain   string `json:"domain" gorm:"unique"`
 }
