@@ -1,49 +1,56 @@
 package model
 
 import (
-	"fmt"
-	"html/template"
 	"log"
-	"sync"
+	"slices"
 	"time"
 
-	"github.com/naiba/nezha/pkg/utils"
-	pb "github.com/naiba/nezha/proto"
 	"gorm.io/gorm"
+
+	"github.com/nezhahq/nezha/pkg/utils"
+	pb "github.com/nezhahq/nezha/proto"
 )
 
 type Server struct {
 	Common
-	Name         string
-	Tag          string   // 分组名
-	Secret       string   `gorm:"uniqueIndex" json:"-"`
-	Note         string   `json:"-"`                    // 管理员可见备注
-	PublicNote   string   `json:"PublicNote,omitempty"` // 公开备注
-	DisplayIndex int      // 展示排序，越大越靠前
-	HideForGuest bool     // 对游客隐藏
-	EnableDDNS   bool     // 启用DDNS
-	DDNSProfiles []uint64 `gorm:"-" json:"-"` // DDNS配置
 
-	DDNSProfilesRaw string `gorm:"default:'[]';column:ddns_profiles_raw" json:"-"`
+	Name                   string `json:"name"`
+	UUID                   string `json:"uuid,omitempty" gorm:"unique"`
+	Note                   string `json:"note,omitempty"`           // 管理员可见备注
+	PublicNote             string `json:"public_note,omitempty"`    // 公开备注
+	DisplayIndex           int    `json:"display_index"`            // 展示排序，越大越靠前
+	HideForGuest           bool   `json:"hide_for_guest,omitempty"` // 对游客隐藏
+	EnableDDNS             bool   `json:"enable_ddns,omitempty"`    // 启用DDNS
+	DDNSProfilesRaw        string `gorm:"default:'[]';column:ddns_profiles_raw" json:"-"`
+	OverrideDDNSDomainsRaw string `gorm:"default:'{}';column:override_ddns_domains_raw" json:"-"`
 
-	Host       *Host      `gorm:"-"`
-	State      *HostState `gorm:"-"`
-	LastActive time.Time  `gorm:"-"`
+	DDNSProfiles        []uint64            `gorm:"-" json:"ddns_profiles,omitempty" validate:"optional"` // DDNS配置
+	OverrideDDNSDomains map[uint64][]string `gorm:"-" json:"override_ddns_domains,omitempty" validate:"optional"`
 
-	TaskClose     chan error                        `gorm:"-" json:"-"`
-	TaskCloseLock *sync.Mutex                       `gorm:"-" json:"-"`
-	TaskStream    pb.NezhaService_RequestTaskServer `gorm:"-" json:"-"`
+	Host       *Host      `gorm:"-" json:"host,omitempty"`
+	State      *HostState `gorm:"-" json:"state,omitempty"`
+	GeoIP      *GeoIP     `gorm:"-" json:"geoip,omitempty"`
+	LastActive time.Time  `gorm:"-" json:"last_active,omitempty"`
+
+	TaskStream  pb.NezhaService_RequestTaskServer `gorm:"-" json:"-"`
+	ConfigCache chan any                          `gorm:"-" json:"-"`
 
 	PrevTransferInSnapshot  int64 `gorm:"-" json:"-"` // 上次数据点时的入站使用量
 	PrevTransferOutSnapshot int64 `gorm:"-" json:"-"` // 上次数据点时的出站使用量
 }
 
+func InitServer(s *Server) {
+	s.Host = &Host{}
+	s.State = &HostState{}
+	s.GeoIP = &GeoIP{}
+	s.ConfigCache = make(chan any, 1)
+}
+
 func (s *Server) CopyFromRunningServer(old *Server) {
 	s.Host = old.Host
 	s.State = old.State
+	s.GeoIP = old.GeoIP
 	s.LastActive = old.LastActive
-	s.TaskClose = old.TaskClose
-	s.TaskCloseLock = old.TaskCloseLock
 	s.TaskStream = old.TaskStream
 	s.PrevTransferInSnapshot = old.PrevTransferInSnapshot
 	s.PrevTransferOutSnapshot = old.PrevTransferOutSnapshot
@@ -56,22 +63,28 @@ func (s *Server) AfterFind(tx *gorm.DB) error {
 			return nil
 		}
 	}
+	if s.OverrideDDNSDomainsRaw != "" {
+		if err := utils.Json.Unmarshal([]byte(s.OverrideDDNSDomainsRaw), &s.OverrideDDNSDomains); err != nil {
+			log.Println("NEZHA>> Server.AfterFind:", err)
+			return nil
+		}
+	}
 	return nil
 }
 
-func boolToString(b bool) string {
-	if b {
-		return "true"
+// Split a sorted server list into two separate lists:
+// The first list contains servers with a priority set (DisplayIndex != 0).
+// The second list contains servers without a priority set (DisplayIndex == 0).
+// The original slice is not modified. If no server without a priority is found, it returns nil.
+func SplitList(x []*Server) ([]*Server, []*Server) {
+	pri := func(s *Server) bool {
+		return s.DisplayIndex == 0
 	}
-	return "false"
-}
 
-func (s Server) MarshalForDashboard() template.JS {
-	name, _ := utils.Json.Marshal(s.Name)
-	tag, _ := utils.Json.Marshal(s.Tag)
-	note, _ := utils.Json.Marshal(s.Note)
-	secret, _ := utils.Json.Marshal(s.Secret)
-	ddnsProfilesRaw, _ := utils.Json.Marshal(s.DDNSProfilesRaw)
-	publicNote, _ := utils.Json.Marshal(s.PublicNote)
-	return template.JS(fmt.Sprintf(`{"ID":%d,"Name":%s,"Secret":%s,"DisplayIndex":%d,"Tag":%s,"Note":%s,"HideForGuest": %s,"EnableDDNS": %s,"DDNSProfilesRaw": %s,"PublicNote": %s}`, s.ID, name, secret, s.DisplayIndex, tag, note, boolToString(s.HideForGuest), boolToString(s.EnableDDNS), ddnsProfilesRaw, publicNote))
+	i := slices.IndexFunc(x, pri)
+	if i == -1 {
+		return nil, x
+	}
+
+	return x[:i], x[i:]
 }
