@@ -14,12 +14,12 @@ var (
 func init() {
 	HttpClientSkipTlsVerify = httpClient(_httpClient{
 		Transport: httpTransport(_httpTransport{
-			SkipVerifySSL: true,
+			SkipVerifyTLS: true,
 		}),
 	})
 	HttpClient = httpClient(_httpClient{
 		Transport: httpTransport(_httpTransport{
-			SkipVerifySSL: false,
+			SkipVerifyTLS: false,
 		}),
 	})
 
@@ -27,12 +27,12 @@ func init() {
 }
 
 type _httpTransport struct {
-	SkipVerifySSL bool
+	SkipVerifyTLS bool
 }
 
 func httpTransport(conf _httpTransport) *http.Transport {
 	return &http.Transport{
-		TLSClientConfig: &tls.Config{InsecureSkipVerify: conf.SkipVerifySSL},
+		TLSClientConfig: &tls.Config{InsecureSkipVerify: conf.SkipVerifyTLS},
 		Proxy:           http.ProxyFromEnvironment,
 	}
 }
