@@ -1,7 +1,7 @@
 // Code generated by protoc-gen-go-grpc. DO NOT EDIT.
 // versions:
 // - protoc-gen-go-grpc v1.3.0
-// - protoc             v5.26.1
+// - protoc             v5.29.3
 // source: proto/nezha.proto
 
 package proto
@@ -21,22 +21,22 @@ const _ = grpc.SupportPackageIsVersion7
 const (
 	NezhaService_ReportSystemState_FullMethodName = "/proto.NezhaService/ReportSystemState"
 	NezhaService_ReportSystemInfo_FullMethodName  = "/proto.NezhaService/ReportSystemInfo"
-	NezhaService_ReportTask_FullMethodName        = "/proto.NezhaService/ReportTask"
 	NezhaService_RequestTask_FullMethodName       = "/proto.NezhaService/RequestTask"
 	NezhaService_IOStream_FullMethodName          = "/proto.NezhaService/IOStream"
-	NezhaService_LookupGeoIP_FullMethodName       = "/proto.NezhaService/LookupGeoIP"
+	NezhaService_ReportGeoIP_FullMethodName       = "/proto.NezhaService/ReportGeoIP"
+	NezhaService_ReportSystemInfo2_FullMethodName = "/proto.NezhaService/ReportSystemInfo2"
 )
 
 // NezhaServiceClient is the client API for NezhaService service.
 //
 // For semantics around ctx use and closing/ending streaming RPCs, please refer to https://pkg.go.dev/google.golang.org/grpc/?tab=doc#ClientConn.NewStream.
 type NezhaServiceClient interface {
-	ReportSystemState(ctx context.Context, in *State, opts ...grpc.CallOption) (*Receipt, error)
+	ReportSystemState(ctx context.Context, opts ...grpc.CallOption) (NezhaService_ReportSystemStateClient, error)
 	ReportSystemInfo(ctx context.Context, in *Host, opts ...grpc.CallOption) (*Receipt, error)
-	ReportTask(ctx context.Context, in *TaskResult, opts ...grpc.CallOption) (*Receipt, error)
-	RequestTask(ctx context.Context, in *Host, opts ...grpc.CallOption) (NezhaService_RequestTaskClient, error)
+	RequestTask(ctx context.Context, opts ...grpc.CallOption) (NezhaService_RequestTaskClient, error)
 	IOStream(ctx context.Context, opts ...grpc.CallOption) (NezhaService_IOStreamClient, error)
-	LookupGeoIP(ctx context.Context, in *GeoIP, opts ...grpc.CallOption) (*GeoIP, error)
+	ReportGeoIP(ctx context.Context, in *GeoIP, opts ...grpc.CallOption) (*GeoIP, error)
+	ReportSystemInfo2(ctx context.Context, in *Host, opts ...grpc.CallOption) (*Uint64Receipt, error)
 }
 
 type nezhaServiceClient struct {
@@ -47,13 +47,35 @@ func NewNezhaServiceClient(cc grpc.ClientConnInterface) NezhaServiceClient {
 	return &nezhaServiceClient{cc}
 }
 
-func (c *nezhaServiceClient) ReportSystemState(ctx context.Context, in *State, opts ...grpc.CallOption) (*Receipt, error) {
-	out := new(Receipt)
-	err := c.cc.Invoke(ctx, NezhaService_ReportSystemState_FullMethodName, in, out, opts...)
+func (c *nezhaServiceClient) ReportSystemState(ctx context.Context, opts ...grpc.CallOption) (NezhaService_ReportSystemStateClient, error) {
+	stream, err := c.cc.NewStream(ctx, &NezhaService_ServiceDesc.Streams[0], NezhaService_ReportSystemState_FullMethodName, opts...)
 	if err != nil {
 		return nil, err
 	}
-	return out, nil
+	x := &nezhaServiceReportSystemStateClient{stream}
+	return x, nil
+}
+
+type NezhaService_ReportSystemStateClient interface {
+	Send(*State) error
+	Recv() (*Receipt, error)
+	grpc.ClientStream
+}
+
+type nezhaServiceReportSystemStateClient struct {
+	grpc.ClientStream
+}
+
+func (x *nezhaServiceReportSystemStateClient) Send(m *State) error {
+	return x.ClientStream.SendMsg(m)
+}
+
+func (x *nezhaServiceReportSystemStateClient) Recv() (*Receipt, error) {
+	m := new(Receipt)
+	if err := x.ClientStream.RecvMsg(m); err != nil {
+		return nil, err
+	}
+	return m, nil
 }
 
 func (c *nezhaServiceClient) ReportSystemInfo(ctx context.Context, in *Host, opts ...grpc.CallOption) (*Receipt, error) {
@@ -65,37 +87,27 @@ func (c *nezhaServiceClient) ReportSystemInfo(ctx context.Context, in *Host, opt
 	return out, nil
 }
 
-func (c *nezhaServiceClient) ReportTask(ctx context.Context, in *TaskResult, opts ...grpc.CallOption) (*Receipt, error) {
-	out := new(Receipt)
-	err := c.cc.Invoke(ctx, NezhaService_ReportTask_FullMethodName, in, out, opts...)
-	if err != nil {
-		return nil, err
-	}
-	return out, nil
-}
-
-func (c *nezhaServiceClient) RequestTask(ctx context.Context, in *Host, opts ...grpc.CallOption) (NezhaService_RequestTaskClient, error) {
-	stream, err := c.cc.NewStream(ctx, &NezhaService_ServiceDesc.Streams[0], NezhaService_RequestTask_FullMethodName, opts...)
+func (c *nezhaServiceClient) RequestTask(ctx context.Context, opts ...grpc.CallOption) (NezhaService_RequestTaskClient, error) {
+	stream, err := c.cc.NewStream(ctx, &NezhaService_ServiceDesc.Streams[1], NezhaService_RequestTask_FullMethodName, opts...)
 	if err != nil {
 		return nil, err
 	}
 	x := &nezhaServiceRequestTaskClient{stream}
-	if err := x.ClientStream.SendMsg(in); err != nil {
-		return nil, err
-	}
-	if err := x.ClientStream.CloseSend(); err != nil {
-		return nil, err
-	}
 	return x, nil
 }
 
 type NezhaService_RequestTaskClient interface {
+	Send(*TaskResult) error
 	Recv() (*Task, error)
 	grpc.ClientStream
 }
 
 type nezhaServiceRequestTaskClient struct {
 	grpc.ClientStream
+}
+
+func (x *nezhaServiceRequestTaskClient) Send(m *TaskResult) error {
+	return x.ClientStream.SendMsg(m)
 }
 
 func (x *nezhaServiceRequestTaskClient) Recv() (*Task, error) {
@@ -107,7 +119,7 @@ func (x *nezhaServiceRequestTaskClient) Recv() (*Task, error) {
 }
 
 func (c *nezhaServiceClient) IOStream(ctx context.Context, opts ...grpc.CallOption) (NezhaService_IOStreamClient, error) {
-	stream, err := c.cc.NewStream(ctx, &NezhaService_ServiceDesc.Streams[1], NezhaService_IOStream_FullMethodName, opts...)
+	stream, err := c.cc.NewStream(ctx, &NezhaService_ServiceDesc.Streams[2], NezhaService_IOStream_FullMethodName, opts...)
 	if err != nil {
 		return nil, err
 	}
@@ -137,9 +149,18 @@ func (x *nezhaServiceIOStreamClient) Recv() (*IOStreamData, error) {
 	return m, nil
 }
 
-func (c *nezhaServiceClient) LookupGeoIP(ctx context.Context, in *GeoIP, opts ...grpc.CallOption) (*GeoIP, error) {
+func (c *nezhaServiceClient) ReportGeoIP(ctx context.Context, in *GeoIP, opts ...grpc.CallOption) (*GeoIP, error) {
 	out := new(GeoIP)
-	err := c.cc.Invoke(ctx, NezhaService_LookupGeoIP_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, NezhaService_ReportGeoIP_FullMethodName, in, out, opts...)
+	if err != nil {
+		return nil, err
+	}
+	return out, nil
+}
+
+func (c *nezhaServiceClient) ReportSystemInfo2(ctx context.Context, in *Host, opts ...grpc.CallOption) (*Uint64Receipt, error) {
+	out := new(Uint64Receipt)
+	err := c.cc.Invoke(ctx, NezhaService_ReportSystemInfo2_FullMethodName, in, out, opts...)
 	if err != nil {
 		return nil, err
 	}
@@ -150,35 +171,35 @@ func (c *nezhaServiceClient) LookupGeoIP(ctx context.Context, in *GeoIP, opts ..
 // All implementations should embed UnimplementedNezhaServiceServer
 // for forward compatibility
 type NezhaServiceServer interface {
-	ReportSystemState(context.Context, *State) (*Receipt, error)
+	ReportSystemState(NezhaService_ReportSystemStateServer) error
 	ReportSystemInfo(context.Context, *Host) (*Receipt, error)
-	ReportTask(context.Context, *TaskResult) (*Receipt, error)
-	RequestTask(*Host, NezhaService_RequestTaskServer) error
+	RequestTask(NezhaService_RequestTaskServer) error
 	IOStream(NezhaService_IOStreamServer) error
-	LookupGeoIP(context.Context, *GeoIP) (*GeoIP, error)
+	ReportGeoIP(context.Context, *GeoIP) (*GeoIP, error)
+	ReportSystemInfo2(context.Context, *Host) (*Uint64Receipt, error)
 }
 
 // UnimplementedNezhaServiceServer should be embedded to have forward compatible implementations.
 type UnimplementedNezhaServiceServer struct {
 }
 
-func (UnimplementedNezhaServiceServer) ReportSystemState(context.Context, *State) (*Receipt, error) {
-	return nil, status.Errorf(codes.Unimplemented, "method ReportSystemState not implemented")
+func (UnimplementedNezhaServiceServer) ReportSystemState(NezhaService_ReportSystemStateServer) error {
+	return status.Errorf(codes.Unimplemented, "method ReportSystemState not implemented")
 }
 func (UnimplementedNezhaServiceServer) ReportSystemInfo(context.Context, *Host) (*Receipt, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method ReportSystemInfo not implemented")
 }
-func (UnimplementedNezhaServiceServer) ReportTask(context.Context, *TaskResult) (*Receipt, error) {
-	return nil, status.Errorf(codes.Unimplemented, "method ReportTask not implemented")
-}
-func (UnimplementedNezhaServiceServer) RequestTask(*Host, NezhaService_RequestTaskServer) error {
+func (UnimplementedNezhaServiceServer) RequestTask(NezhaService_RequestTaskServer) error {
 	return status.Errorf(codes.Unimplemented, "method RequestTask not implemented")
 }
 func (UnimplementedNezhaServiceServer) IOStream(NezhaService_IOStreamServer) error {
 	return status.Errorf(codes.Unimplemented, "method IOStream not implemented")
 }
-func (UnimplementedNezhaServiceServer) LookupGeoIP(context.Context, *GeoIP) (*GeoIP, error) {
-	return nil, status.Errorf(codes.Unimplemented, "method LookupGeoIP not implemented")
+func (UnimplementedNezhaServiceServer) ReportGeoIP(context.Context, *GeoIP) (*GeoIP, error) {
+	return nil, status.Errorf(codes.Unimplemented, "method ReportGeoIP not implemented")
+}
+func (UnimplementedNezhaServiceServer) ReportSystemInfo2(context.Context, *Host) (*Uint64Receipt, error) {
+	return nil, status.Errorf(codes.Unimplemented, "method ReportSystemInfo2 not implemented")
 }
 
 // UnsafeNezhaServiceServer may be embedded to opt out of forward compatibility for this service.
@@ -192,22 +213,30 @@ func RegisterNezhaServiceServer(s grpc.ServiceRegistrar, srv NezhaServiceServer)
 	s.RegisterService(&NezhaService_ServiceDesc, srv)
 }
 
-func _NezhaService_ReportSystemState_Handler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (interface{}, error) {
-	in := new(State)
-	if err := dec(in); err != nil {
+func _NezhaService_ReportSystemState_Handler(srv interface{}, stream grpc.ServerStream) error {
+	return srv.(NezhaServiceServer).ReportSystemState(&nezhaServiceReportSystemStateServer{stream})
+}
+
+type NezhaService_ReportSystemStateServer interface {
+	Send(*Receipt) error
+	Recv() (*State, error)
+	grpc.ServerStream
+}
+
+type nezhaServiceReportSystemStateServer struct {
+	grpc.ServerStream
+}
+
+func (x *nezhaServiceReportSystemStateServer) Send(m *Receipt) error {
+	return x.ServerStream.SendMsg(m)
+}
+
+func (x *nezhaServiceReportSystemStateServer) Recv() (*State, error) {
+	m := new(State)
+	if err := x.ServerStream.RecvMsg(m); err != nil {
 		return nil, err
 	}
-	if interceptor == nil {
-		return srv.(NezhaServiceServer).ReportSystemState(ctx, in)
-	}
-	info := &grpc.UnaryServerInfo{
-		Server:     srv,
-		FullMethod: NezhaService_ReportSystemState_FullMethodName,
-	}
-	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
-		return srv.(NezhaServiceServer).ReportSystemState(ctx, req.(*State))
-	}
-	return interceptor(ctx, in, info, handler)
+	return m, nil
 }
 
 func _NezhaService_ReportSystemInfo_Handler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (interface{}, error) {
@@ -228,34 +257,13 @@ func _NezhaService_ReportSystemInfo_Handler(srv interface{}, ctx context.Context
 	return interceptor(ctx, in, info, handler)
 }
 
-func _NezhaService_ReportTask_Handler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (interface{}, error) {
-	in := new(TaskResult)
-	if err := dec(in); err != nil {
-		return nil, err
-	}
-	if interceptor == nil {
-		return srv.(NezhaServiceServer).ReportTask(ctx, in)
-	}
-	info := &grpc.UnaryServerInfo{
-		Server:     srv,
-		FullMethod: NezhaService_ReportTask_FullMethodName,
-	}
-	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
-		return srv.(NezhaServiceServer).ReportTask(ctx, req.(*TaskResult))
-	}
-	return interceptor(ctx, in, info, handler)
-}
-
 func _NezhaService_RequestTask_Handler(srv interface{}, stream grpc.ServerStream) error {
-	m := new(Host)
-	if err := stream.RecvMsg(m); err != nil {
-		return err
-	}
-	return srv.(NezhaServiceServer).RequestTask(m, &nezhaServiceRequestTaskServer{stream})
+	return srv.(NezhaServiceServer).RequestTask(&nezhaServiceRequestTaskServer{stream})
 }
 
 type NezhaService_RequestTaskServer interface {
 	Send(*Task) error
+	Recv() (*TaskResult, error)
 	grpc.ServerStream
 }
 
@@ -265,6 +273,14 @@ type nezhaServiceRequestTaskServer struct {
 
 func (x *nezhaServiceRequestTaskServer) Send(m *Task) error {
 	return x.ServerStream.SendMsg(m)
+}
+
+func (x *nezhaServiceRequestTaskServer) Recv() (*TaskResult, error) {
+	m := new(TaskResult)
+	if err := x.ServerStream.RecvMsg(m); err != nil {
+		return nil, err
+	}
+	return m, nil
 }
 
 func _NezhaService_IOStream_Handler(srv interface{}, stream grpc.ServerStream) error {
@@ -293,20 +309,38 @@ func (x *nezhaServiceIOStreamServer) Recv() (*IOStreamData, error) {
 	return m, nil
 }
 
-func _NezhaService_LookupGeoIP_Handler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (interface{}, error) {
+func _NezhaService_ReportGeoIP_Handler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (interface{}, error) {
 	in := new(GeoIP)
 	if err := dec(in); err != nil {
 		return nil, err
 	}
 	if interceptor == nil {
-		return srv.(NezhaServiceServer).LookupGeoIP(ctx, in)
+		return srv.(NezhaServiceServer).ReportGeoIP(ctx, in)
 	}
 	info := &grpc.UnaryServerInfo{
 		Server:     srv,
-		FullMethod: NezhaService_LookupGeoIP_FullMethodName,
+		FullMethod: NezhaService_ReportGeoIP_FullMethodName,
 	}
 	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
-		return srv.(NezhaServiceServer).LookupGeoIP(ctx, req.(*GeoIP))
+		return srv.(NezhaServiceServer).ReportGeoIP(ctx, req.(*GeoIP))
+	}
+	return interceptor(ctx, in, info, handler)
+}
+
+func _NezhaService_ReportSystemInfo2_Handler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (interface{}, error) {
+	in := new(Host)
+	if err := dec(in); err != nil {
+		return nil, err
+	}
+	if interceptor == nil {
+		return srv.(NezhaServiceServer).ReportSystemInfo2(ctx, in)
+	}
+	info := &grpc.UnaryServerInfo{
+		Server:     srv,
+		FullMethod: NezhaService_ReportSystemInfo2_FullMethodName,
+	}
+	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
+		return srv.(NezhaServiceServer).ReportSystemInfo2(ctx, req.(*Host))
 	}
 	return interceptor(ctx, in, info, handler)
 }
@@ -319,27 +353,30 @@ var NezhaService_ServiceDesc = grpc.ServiceDesc{
 	HandlerType: (*NezhaServiceServer)(nil),
 	Methods: []grpc.MethodDesc{
 		{
-			MethodName: "ReportSystemState",
-			Handler:    _NezhaService_ReportSystemState_Handler,
-		},
-		{
 			MethodName: "ReportSystemInfo",
 			Handler:    _NezhaService_ReportSystemInfo_Handler,
 		},
 		{
-			MethodName: "ReportTask",
-			Handler:    _NezhaService_ReportTask_Handler,
+			MethodName: "ReportGeoIP",
+			Handler:    _NezhaService_ReportGeoIP_Handler,
 		},
 		{
-			MethodName: "LookupGeoIP",
-			Handler:    _NezhaService_LookupGeoIP_Handler,
+			MethodName: "ReportSystemInfo2",
+			Handler:    _NezhaService_ReportSystemInfo2_Handler,
 		},
 	},
 	Streams: []grpc.StreamDesc{
 		{
+			StreamName:    "ReportSystemState",
+			Handler:       _NezhaService_ReportSystemState_Handler,
+			ServerStreams: true,
+			ClientStreams: true,
+		},
+		{
 			StreamName:    "RequestTask",
 			Handler:       _NezhaService_RequestTask_Handler,
 			ServerStreams: true,
+			ClientStreams: true,
 		},
 		{
 			StreamName:    "IOStream",
