@@ -2,24 +2,21 @@ package rpc
 
 import (
 	"context"
+	"strings"
 
+	petname "github.com/dustinkirkland/golang-petname"
+	"github.com/hashicorp/go-uuid"
 	"google.golang.org/grpc/codes"
 	"google.golang.org/grpc/metadata"
 	"google.golang.org/grpc/status"
 
-	"github.com/naiba/nezha/service/singleton"
+	"github.com/nezhahq/nezha/model"
+	"github.com/nezhahq/nezha/service/singleton"
 )
 
 type authHandler struct {
 	ClientSecret string
-}
-
-func (a *authHandler) GetRequestMetadata(ctx context.Context, uri ...string) (map[string]string, error) {
-	return map[string]string{"client_secret": a.ClientSecret}, nil
-}
-
-func (a *authHandler) RequireTransportSecurity() bool {
-	return false
+	ClientUUID   string
 }
 
 func (a *authHandler) Check(ctx context.Context) (uint64, error) {
@@ -30,18 +27,49 @@ func (a *authHandler) Check(ctx context.Context) (uint64, error) {
 
 	var clientSecret string
 	if value, ok := md["client_secret"]; ok {
-		clientSecret = value[0]
+		clientSecret = strings.TrimSpace(value[0])
 	}
 
-	singleton.ServerLock.RLock()
-	defer singleton.ServerLock.RUnlock()
-	clientID, hasID := singleton.SecretToID[clientSecret]
+	if clientSecret == "" {
+		return 0, status.Error(codes.Unauthenticated, "客户端认证失败")
+	}
+
+	ip, _ := ctx.Value(model.CtxKeyRealIP{}).(string)
+
+	singleton.UserLock.RLock()
+	userId, ok := singleton.AgentSecretToUserId[clientSecret]
+	if !ok {
+		singleton.UserLock.RUnlock()
+		model.BlockIP(singleton.DB, ip, model.WAFBlockReasonTypeAgentAuthFail, model.BlockIDgRPC)
+		return 0, status.Error(codes.Unauthenticated, "客户端认证失败")
+	}
+	singleton.UserLock.RUnlock()
+
+	model.UnblockIP(singleton.DB, ip, model.BlockIDgRPC)
+
+	var clientUUID string
+	if value, ok := md["client_uuid"]; ok {
+		clientUUID = value[0]
+	}
+
+	if _, err := uuid.ParseUUID(clientUUID); err != nil {
+		return 0, status.Error(codes.Unauthenticated, "客户端 UUID 不合法")
+	}
+
+	clientID, hasID := singleton.ServerShared.UUIDToID(clientUUID)
 	if !hasID {
-		return 0, status.Errorf(codes.Unauthenticated, "客户端认证失败")
+		s := model.Server{UUID: clientUUID, Name: petname.Generate(2, "-"), Common: model.Common{
+			UserID: userId,
+		}}
+		if err := singleton.DB.Create(&s).Error; err != nil {
+			return 0, status.Error(codes.Unauthenticated, err.Error())
+		}
+
+		model.InitServer(&s)
+		singleton.ServerShared.Update(&s, clientUUID)
+
+		clientID = s.ID
 	}
-	_, hasServer := singleton.ServerList[clientID]
-	if !hasServer {
-		return 0, status.Errorf(codes.Unauthenticated, "客户端认证失败")
-	}
+
 	return clientID, nil
 }
