@@ -2,24 +2,24 @@ package rpc
 
 import (
 	"context"
+	"errors"
 	"fmt"
 	"log"
 	"net"
 	"sync"
 	"time"
 
-	"github.com/naiba/nezha/pkg/ddns"
-	"github.com/naiba/nezha/pkg/geoip"
-	"github.com/naiba/nezha/pkg/grpcx"
-	"github.com/naiba/nezha/pkg/utils"
-
 	"github.com/jinzhu/copier"
-	"github.com/nicksnyder/go-i18n/v2/i18n"
+	"github.com/nezhahq/nezha/pkg/ddns"
+	geoipx "github.com/nezhahq/nezha/pkg/geoip"
+	"github.com/nezhahq/nezha/pkg/grpcx"
 
-	"github.com/naiba/nezha/model"
-	pb "github.com/naiba/nezha/proto"
-	"github.com/naiba/nezha/service/singleton"
+	"github.com/nezhahq/nezha/model"
+	pb "github.com/nezhahq/nezha/proto"
+	"github.com/nezhahq/nezha/service/singleton"
 )
+
+var _ pb.NezhaServiceServer = (*NezhaHandler)(nil)
 
 var NezhaHandlerSingleton *NezhaHandler
 
@@ -37,136 +37,105 @@ func NewNezhaHandler() *NezhaHandler {
 	}
 }
 
-func (s *NezhaHandler) ReportTask(c context.Context, r *pb.TaskResult) (*pb.Receipt, error) {
-	var err error
-	var clientID uint64
-	if clientID, err = s.Auth.Check(c); err != nil {
-		return nil, err
-	}
-	if r.GetType() == model.TaskTypeCommand {
-		// 处理上报的计划任务
-		singleton.CronLock.RLock()
-		defer singleton.CronLock.RUnlock()
-		cr := singleton.Crons[r.GetId()]
-		if cr != nil {
-			singleton.ServerLock.RLock()
-			defer singleton.ServerLock.RUnlock()
-			// 保存当前服务器状态信息
-			curServer := model.Server{}
-			copier.Copy(&curServer, singleton.ServerList[clientID])
-			if cr.PushSuccessful && r.GetSuccessful() {
-				singleton.SendNotification(cr.NotificationTag, fmt.Sprintf("[%s] %s, %s\n%s", singleton.Localizer.MustLocalize(
-					&i18n.LocalizeConfig{
-						MessageID: "ScheduledTaskExecutedSuccessfully",
-					},
-				), cr.Name, singleton.ServerList[clientID].Name, r.GetData()), nil, &curServer)
-			}
-			if !r.GetSuccessful() {
-				singleton.SendNotification(cr.NotificationTag, fmt.Sprintf("[%s] %s, %s\n%s", singleton.Localizer.MustLocalize(
-					&i18n.LocalizeConfig{
-						MessageID: "ScheduledTaskExecutedFailed",
-					},
-				), cr.Name, singleton.ServerList[clientID].Name, r.GetData()), nil, &curServer)
-			}
-			singleton.DB.Model(cr).Updates(model.Cron{
-				LastExecutedAt: time.Now().Add(time.Second * -1 * time.Duration(r.GetDelay())),
-				LastResult:     r.GetSuccessful(),
-			})
-		}
-	} else if model.IsServiceSentinelNeeded(r.GetType()) {
-		singleton.ServiceSentinelShared.Dispatch(singleton.ReportData{
-			Data:     r,
-			Reporter: clientID,
-		})
-	}
-	return &pb.Receipt{Proced: true}, nil
-}
-
-func (s *NezhaHandler) RequestTask(h *pb.Host, stream pb.NezhaService_RequestTaskServer) error {
+func (s *NezhaHandler) RequestTask(stream pb.NezhaService_RequestTaskServer) error {
 	var clientID uint64
 	var err error
 	if clientID, err = s.Auth.Check(stream.Context()); err != nil {
 		return err
 	}
-	closeCh := make(chan error)
-	singleton.ServerLock.RLock()
-	singleton.ServerList[clientID].TaskCloseLock.Lock()
-	// 修复不断的请求 task 但是没有 return 导致内存泄漏
-	if singleton.ServerList[clientID].TaskClose != nil {
-		close(singleton.ServerList[clientID].TaskClose)
-	}
-	singleton.ServerList[clientID].TaskStream = stream
-	singleton.ServerList[clientID].TaskClose = closeCh
-	singleton.ServerList[clientID].TaskCloseLock.Unlock()
-	singleton.ServerLock.RUnlock()
-	return <-closeCh
-}
 
-func (s *NezhaHandler) ReportSystemState(c context.Context, r *pb.State) (*pb.Receipt, error) {
-	var clientID uint64
-	var err error
-	if clientID, err = s.Auth.Check(c); err != nil {
-		return nil, err
-	}
-	state := model.PB2State(r)
-	singleton.ServerLock.RLock()
-	defer singleton.ServerLock.RUnlock()
-	singleton.ServerList[clientID].LastActive = time.Now()
-	singleton.ServerList[clientID].State = &state
-
-	// 应对 dashboard 重启的情况，如果从未记录过，先打点，等到小时时间点时入库
-	if singleton.ServerList[clientID].PrevTransferInSnapshot == 0 || singleton.ServerList[clientID].PrevTransferOutSnapshot == 0 {
-		singleton.ServerList[clientID].PrevTransferInSnapshot = int64(state.NetInTransfer)
-		singleton.ServerList[clientID].PrevTransferOutSnapshot = int64(state.NetOutTransfer)
-	}
-
-	return &pb.Receipt{Proced: true}, nil
-}
-
-func (s *NezhaHandler) ReportSystemInfo(c context.Context, r *pb.Host) (*pb.Receipt, error) {
-	var clientID uint64
-	var err error
-	if clientID, err = s.Auth.Check(c); err != nil {
-		return nil, err
-	}
-	host := model.PB2Host(r)
-	singleton.ServerLock.RLock()
-	defer singleton.ServerLock.RUnlock()
-
-	// 检查并更新DDNS
-	if singleton.ServerList[clientID].EnableDDNS && host.IP != "" &&
-		(singleton.ServerList[clientID].Host == nil || singleton.ServerList[clientID].Host.IP != host.IP) {
-		ipv4, ipv6, _ := utils.SplitIPAddr(host.IP)
-		providers, err := singleton.GetDDNSProvidersFromProfiles(singleton.ServerList[clientID].DDNSProfiles, &ddns.IP{Ipv4Addr: ipv4, Ipv6Addr: ipv6})
-		if err == nil {
-			for _, provider := range providers {
-				go func(provider *ddns.Provider) {
-					provider.UpdateDomain(context.Background())
-				}(provider)
+	server, _ := singleton.ServerShared.Get(clientID)
+	server.TaskStream = stream
+	var result *pb.TaskResult
+	for {
+		result, err = stream.Recv()
+		if err != nil {
+			log.Printf("NEZHA>> RequestTask error: %v, clientID: %d\n", err, clientID)
+			return nil
+		}
+		switch result.GetType() {
+		case model.TaskTypeCommand:
+			// 处理上报的计划任务
+			cr, _ := singleton.CronShared.Get(result.GetId())
+			if cr != nil {
+				// 保存当前服务器状态信息
+				var curServer model.Server
+				copier.Copy(&curServer, server)
+				if cr.PushSuccessful && result.GetSuccessful() {
+					singleton.NotificationShared.SendNotification(cr.NotificationGroupID, fmt.Sprintf("[%s] %s, %s\n%s", singleton.Localizer.T("Scheduled Task Executed Successfully"),
+						cr.Name, server.Name, result.GetData()), nil, &curServer)
+				}
+				if !result.GetSuccessful() {
+					singleton.NotificationShared.SendNotification(cr.NotificationGroupID, fmt.Sprintf("[%s] %s, %s\n%s", singleton.Localizer.T("Scheduled Task Executed Failed"),
+						cr.Name, server.Name, result.GetData()), nil, &curServer)
+				}
+				singleton.DB.Model(cr).Updates(model.Cron{
+					LastExecutedAt: time.Now().Add(time.Second * -1 * time.Duration(result.GetDelay())),
+					LastResult:     result.GetSuccessful(),
+				})
 			}
-		} else {
-			log.Printf("NEZHA>> 获取DDNS配置时发生错误: %v", err)
+		case model.TaskTypeReportConfig:
+			if len(server.ConfigCache) < 1 {
+				if !result.GetSuccessful() {
+					server.ConfigCache <- errors.New(result.Data)
+					continue
+				}
+				server.ConfigCache <- result.Data
+			}
+		default:
+			if model.IsServiceSentinelNeeded(result.GetType()) {
+				singleton.ServiceSentinelShared.Dispatch(singleton.ReportData{
+					Data:     result,
+					Reporter: clientID,
+				})
+			}
 		}
 	}
+}
 
-	// 发送IP变动通知
-	if singleton.ServerList[clientID].Host != nil && singleton.Conf.EnableIPChangeNotification &&
-		((singleton.Conf.Cover == model.ConfigCoverAll && !singleton.Conf.IgnoredIPNotificationServerIDs[clientID]) ||
-			(singleton.Conf.Cover == model.ConfigCoverIgnoreAll && singleton.Conf.IgnoredIPNotificationServerIDs[clientID])) &&
-		singleton.ServerList[clientID].Host.IP != "" &&
-		host.IP != "" &&
-		singleton.ServerList[clientID].Host.IP != host.IP {
+func (s *NezhaHandler) ReportSystemState(stream pb.NezhaService_ReportSystemStateServer) error {
+	var err error
+	var clientID uint64
+	if clientID, err = s.Auth.Check(stream.Context()); err != nil {
+		return err
+	}
+	var state *pb.State
+	for {
+		state, err = stream.Recv()
+		if err != nil {
+			log.Printf("NEZHA>> ReportSystemState eror: %v, clientID: %d\n", err, clientID)
+			return nil
+		}
+		state := model.PB2State(state)
 
-		singleton.SendNotification(singleton.Conf.IPChangeNotificationTag,
-			fmt.Sprintf(
-				"[%s] %s, %s => %s",
-				singleton.Localizer.MustLocalize(&i18n.LocalizeConfig{
-					MessageID: "IPChanged",
-				}),
-				singleton.ServerList[clientID].Name, singleton.IPDesensitize(singleton.ServerList[clientID].Host.IP),
-				singleton.IPDesensitize(host.IP),
-			),
-			nil)
+		server, ok := singleton.ServerShared.Get(clientID)
+		if !ok || server == nil {
+			return nil
+		}
+
+		server.LastActive = time.Now()
+		server.State = &state
+		// 应对 dashboard 重启的情况，如果从未记录过，先打点，等到小时时间点时入库
+		if server.PrevTransferInSnapshot == 0 || server.PrevTransferOutSnapshot == 0 {
+			server.PrevTransferInSnapshot = int64(state.NetInTransfer)
+			server.PrevTransferOutSnapshot = int64(state.NetOutTransfer)
+		}
+
+		stream.Send(&pb.Receipt{Proced: true})
+	}
+}
+
+func (s *NezhaHandler) onReportSystemInfo(c context.Context, r *pb.Host) error {
+	var clientID uint64
+	var err error
+	if clientID, err = s.Auth.Check(c); err != nil {
+		return err
+	}
+	host := model.PB2Host(r)
+
+	server, ok := singleton.ServerShared.Get(clientID)
+	if !ok || server == nil {
+		return fmt.Errorf("server not found")
 	}
 
 	/**
@@ -174,18 +143,23 @@ func (s *NezhaHandler) ReportSystemInfo(c context.Context, r *pb.Host) (*pb.Rece
 	 * 当 agent 重启时，bootTime 变大，agent 端会先上报 host 信息，然后上报 state 信息
 	 * 这是可以借助上报顺序的空档，将停机前的流量统计数据标记下来，加到下一个小时的数据点上
 	 */
-	if singleton.ServerList[clientID].Host != nil && singleton.ServerList[clientID].Host.BootTime < host.BootTime {
-		singleton.ServerList[clientID].PrevTransferInSnapshot = singleton.ServerList[clientID].PrevTransferInSnapshot - int64(singleton.ServerList[clientID].State.NetInTransfer)
-		singleton.ServerList[clientID].PrevTransferOutSnapshot = singleton.ServerList[clientID].PrevTransferOutSnapshot - int64(singleton.ServerList[clientID].State.NetOutTransfer)
+	if server.Host != nil && server.Host.BootTime < host.BootTime {
+		server.PrevTransferInSnapshot = server.PrevTransferInSnapshot - int64(server.State.NetInTransfer)
+		server.PrevTransferOutSnapshot = server.PrevTransferOutSnapshot - int64(server.State.NetOutTransfer)
 	}
 
-	// 不要冲掉国家码
-	if singleton.ServerList[clientID].Host != nil {
-		host.CountryCode = singleton.ServerList[clientID].Host.CountryCode
-	}
+	server.Host = &host
+	return nil
+}
 
-	singleton.ServerList[clientID].Host = &host
+func (s *NezhaHandler) ReportSystemInfo(c context.Context, r *pb.Host) (*pb.Receipt, error) {
+	s.onReportSystemInfo(c, r)
 	return &pb.Receipt{Proced: true}, nil
+}
+
+func (s *NezhaHandler) ReportSystemInfo2(c context.Context, r *pb.Host) (*pb.Uint64Receipt, error) {
+	s.onReportSystemInfo(c, r)
+	return &pb.Uint64Receipt{Data: singleton.DashboardBootTime}, nil
 }
 
 func (s *NezhaHandler) IOStream(stream pb.NezhaService_IOStreamServer) error {
@@ -196,9 +170,21 @@ func (s *NezhaHandler) IOStream(stream pb.NezhaService_IOStreamServer) error {
 	if err != nil {
 		return err
 	}
+
+	// ff05ff05 是 Nezha 的魔数，用于标识流 ID
 	if id == nil || len(id.Data) < 4 || (id.Data[0] != 0xff && id.Data[1] != 0x05 && id.Data[2] != 0xff && id.Data[3] == 0x05) {
 		return fmt.Errorf("invalid stream id")
 	}
+
+	go func() {
+		for {
+			if err := stream.Send(&pb.IOStreamData{Data: []byte{}}); err != nil {
+				log.Printf("NEZHA>> IOStream keepAlive error: %v\n", err)
+				return
+			}
+			time.Sleep(time.Second * 30)
+		}
+	}()
 
 	streamId := string(id.Data[4:])
 
@@ -213,29 +199,85 @@ func (s *NezhaHandler) IOStream(stream pb.NezhaService_IOStreamServer) error {
 	return nil
 }
 
-func (s *NezhaHandler) LookupGeoIP(c context.Context, r *pb.GeoIP) (*pb.GeoIP, error) {
+func (s *NezhaHandler) ReportGeoIP(c context.Context, r *pb.GeoIP) (*pb.GeoIP, error) {
 	var clientID uint64
 	var err error
 	if clientID, err = s.Auth.Check(c); err != nil {
 		return nil, err
 	}
 
-	// 根据内置数据库查询 IP 地理位置
-	record := &geoip.IPInfo{}
-	ip := r.GetIp()
-	netIP := net.ParseIP(ip)
-	location, err := geoip.Lookup(netIP, record)
-	if err != nil {
-		return nil, err
+	geoip := model.PB2GeoIP(r)
+	use6 := r.GetUse6()
+
+	if geoip.IP.IPv4Addr == "" && geoip.IP.IPv6Addr == "" {
+		ip, _ := c.Value(model.CtxKeyRealIP{}).(string)
+		if ip == "" {
+			ip, _ = c.Value(model.CtxKeyConnectingIP{}).(string)
+		}
+		geoip.IP.IPv4Addr = ip
 	}
+
+	joinedIP := geoip.IP.Join()
+
+	server, ok := singleton.ServerShared.Get(clientID)
+	if !ok || server == nil {
+		return nil, fmt.Errorf("server not found")
+	}
+
+	// 检查并更新DDNS
+	if server.EnableDDNS && joinedIP != "" &&
+		(server.GeoIP == nil || server.GeoIP.IP != geoip.IP) {
+		ipv4 := geoip.IP.IPv4Addr
+		ipv6 := geoip.IP.IPv6Addr
+
+		providers, err := singleton.DDNSShared.GetDDNSProvidersFromProfiles(server.DDNSProfiles, &model.IP{IPv4Addr: ipv4, IPv6Addr: ipv6})
+		if err == nil {
+			for _, provider := range providers {
+				domains := server.OverrideDDNSDomains[provider.GetProfileID()]
+				go func(provider *ddns.Provider) {
+					provider.UpdateDomain(context.Background(), domains...)
+				}(provider)
+			}
+		} else {
+			log.Printf("NEZHA>> Failed to retrieve DDNS configuration: %v", err)
+		}
+	}
+
+	// 发送IP变动通知
+	if server.GeoIP != nil && singleton.Conf.EnableIPChangeNotification &&
+		((singleton.Conf.Cover == model.ConfigCoverAll && !singleton.Conf.IgnoredIPNotificationServerIDs[clientID]) ||
+			(singleton.Conf.Cover == model.ConfigCoverIgnoreAll && singleton.Conf.IgnoredIPNotificationServerIDs[clientID])) &&
+		server.GeoIP.IP.Join() != "" &&
+		joinedIP != "" &&
+		server.GeoIP.IP != geoip.IP {
+
+		singleton.NotificationShared.SendNotification(singleton.Conf.IPChangeNotificationGroupID,
+			fmt.Sprintf(
+				"[%s] %s, %s => %s",
+				singleton.Localizer.T("IP Changed"),
+				server.Name, singleton.IPDesensitize(server.GeoIP.IP.Join()),
+				singleton.IPDesensitize(joinedIP),
+			),
+			nil)
+	}
+
+	// 根据内置数据库查询 IP 地理位置
+	var ip string
+	if geoip.IP.IPv6Addr != "" && (use6 || geoip.IP.IPv4Addr == "") {
+		ip = geoip.IP.IPv6Addr
+	} else {
+		ip = geoip.IP.IPv4Addr
+	}
+
+	netIP := net.ParseIP(ip)
+	location, err := geoipx.Lookup(netIP)
+	if err != nil {
+		log.Printf("NEZHA>> geoip.Lookup: %v", err)
+	}
+	geoip.CountryCode = location
 
 	// 将地区码写入到 Host
-	singleton.ServerLock.RLock()
-	defer singleton.ServerLock.RUnlock()
-	if singleton.ServerList[clientID].Host == nil {
-		return nil, fmt.Errorf("host not found")
-	}
-	singleton.ServerList[clientID].Host.CountryCode = location
+	server.GeoIP = &geoip
 
-	return &pb.GeoIP{Ip: ip, CountryCode: location}, nil
+	return &pb.GeoIP{Ip: nil, CountryCode: location, DashboardBootTime: singleton.DashboardBootTime}, nil
 }
