@@ -1,91 +1,136 @@
 package singleton
 
 import (
-	"bytes"
+	"cmp"
 	"fmt"
-	"sync"
+	"slices"
+	"strings"
 
 	"github.com/jinzhu/copier"
 
 	"github.com/robfig/cron/v3"
 
-	"github.com/naiba/nezha/model"
-	pb "github.com/naiba/nezha/proto"
+	"github.com/nezhahq/nezha/model"
+	"github.com/nezhahq/nezha/pkg/utils"
+	pb "github.com/nezhahq/nezha/proto"
 )
 
-var (
-	Cron     *cron.Cron
-	Crons    map[uint64]*model.Cron // [CrondID] -> *model.Cron
-	CronLock sync.RWMutex
-)
-
-func InitCronTask() {
-	Cron = cron.New(cron.WithSeconds(), cron.WithLocation(Loc))
-	Crons = make(map[uint64]*model.Cron)
+type CronClass struct {
+	class[uint64, *model.Cron]
+	*cron.Cron
 }
 
-// loadCronTasks 加载计划任务
-func loadCronTasks() {
-	InitCronTask()
-	var crons []model.Cron
-	DB.Find(&crons)
+func NewCronClass() *CronClass {
+	cronx := cron.New(cron.WithSeconds(), cron.WithLocation(Loc))
+	list := make(map[uint64]*model.Cron)
+
+	var sortedList []*model.Cron
+	DB.Find(&sortedList)
+
 	var err error
-	var notificationTagList []string
-	notificationMsgMap := make(map[string]*bytes.Buffer)
-	for i := 0; i < len(crons); i++ {
+	var notificationGroupList []uint64
+	notificationMsgMap := make(map[uint64]*strings.Builder)
+
+	for _, cron := range sortedList {
 		// 触发任务类型无需注册
-		if crons[i].TaskType == model.CronTypeTriggerTask {
-			Crons[crons[i].ID] = &crons[i]
+		if cron.TaskType == model.CronTypeTriggerTask {
+			list[cron.ID] = cron
 			continue
 		}
-		// 旧版本计划任务可能不存在通知组 为其添加默认通知组
-		if crons[i].NotificationTag == "" {
-			crons[i].NotificationTag = "default"
-			DB.Save(crons[i])
-		}
 		// 注册计划任务
-		crons[i].CronJobID, err = Cron.AddFunc(crons[i].Scheduler, CronTrigger(crons[i]))
+		cron.CronJobID, err = cronx.AddFunc(cron.Scheduler, CronTrigger(cron))
 		if err == nil {
-			Crons[crons[i].ID] = &crons[i]
+			list[cron.ID] = cron
 		} else {
 			// 当前通知组首次出现 将其加入通知组列表并初始化通知组消息缓存
-			if _, ok := notificationMsgMap[crons[i].NotificationTag]; !ok {
-				notificationTagList = append(notificationTagList, crons[i].NotificationTag)
-				notificationMsgMap[crons[i].NotificationTag] = bytes.NewBufferString("")
-				notificationMsgMap[crons[i].NotificationTag].WriteString("调度失败的计划任务：[")
+			if _, ok := notificationMsgMap[cron.NotificationGroupID]; !ok {
+				notificationGroupList = append(notificationGroupList, cron.NotificationGroupID)
+				notificationMsgMap[cron.NotificationGroupID] = new(strings.Builder)
+				notificationMsgMap[cron.NotificationGroupID].WriteString(Localizer.T("Tasks failed to register: ["))
 			}
-			notificationMsgMap[crons[i].NotificationTag].WriteString(fmt.Sprintf("%d,", crons[i].ID))
+			notificationMsgMap[cron.NotificationGroupID].WriteString(fmt.Sprintf("%d,", cron.ID))
 		}
 	}
+
 	// 向注册错误的计划任务所在通知组发送通知
-	for _, tag := range notificationTagList {
-		notificationMsgMap[tag].WriteString("] 这些任务将无法正常执行,请进入后点重新修改保存。")
-		SendNotification(tag, notificationMsgMap[tag].String(), nil)
+	for _, gid := range notificationGroupList {
+		notificationMsgMap[gid].WriteString(Localizer.T("] These tasks will not execute properly. Fix them in the admin dashboard."))
+		NotificationShared.SendNotification(gid, notificationMsgMap[gid].String(), nil)
 	}
-	Cron.Start()
+	cronx.Start()
+
+	return &CronClass{
+		class: class[uint64, *model.Cron]{
+			list:       list,
+			sortedList: sortedList,
+		},
+		Cron: cronx,
+	}
 }
 
-func ManualTrigger(c model.Cron) {
-	CronTrigger(c)()
+func (c *CronClass) Update(cr *model.Cron) {
+	c.listMu.Lock()
+	crOld := c.list[cr.ID]
+	if crOld != nil && crOld.CronJobID != 0 {
+		c.Cron.Remove(crOld.CronJobID)
+	}
+
+	delete(c.list, cr.ID)
+	c.list[cr.ID] = cr
+	c.listMu.Unlock()
+
+	c.sortList()
 }
 
-func SendTriggerTasks(taskIDs []uint64, triggerServer uint64) {
-	CronLock.RLock()
+func (c *CronClass) Delete(idList []uint64) {
+	c.listMu.Lock()
+	for _, id := range idList {
+		cr := c.list[id]
+		if cr != nil && cr.CronJobID != 0 {
+			c.Cron.Remove(cr.CronJobID)
+		}
+		delete(c.list, id)
+	}
+	c.listMu.Unlock()
+
+	c.sortList()
+}
+
+func (c *CronClass) sortList() {
+	c.listMu.RLock()
+	defer c.listMu.RUnlock()
+
+	sortedList := utils.MapValuesToSlice(c.list)
+	slices.SortFunc(sortedList, func(a, b *model.Cron) int {
+		return cmp.Compare(a.ID, b.ID)
+	})
+
+	c.sortedListMu.Lock()
+	defer c.sortedListMu.Unlock()
+	c.sortedList = sortedList
+}
+
+func (c *CronClass) SendTriggerTasks(taskIDs []uint64, triggerServer uint64) {
+	c.listMu.RLock()
 	var cronLists []*model.Cron
 	for _, taskID := range taskIDs {
-		if c, ok := Crons[taskID]; ok {
+		if c, ok := c.list[taskID]; ok {
 			cronLists = append(cronLists, c)
 		}
 	}
-	CronLock.RUnlock()
+	c.listMu.RUnlock()
 
 	// 依次调用CronTrigger发送任务
 	for _, c := range cronLists {
-		go CronTrigger(*c, triggerServer)()
+		go CronTrigger(c, triggerServer)()
 	}
 }
 
-func CronTrigger(cr model.Cron, triggerServer ...uint64) func() {
+func ManualTrigger(cr *model.Cron) {
+	CronTrigger(cr)()
+}
+
+func CronTrigger(cr *model.Cron, triggerServer ...uint64) func() {
 	crIgnoreMap := make(map[uint64]bool)
 	for j := 0; j < len(cr.Servers); j++ {
 		crIgnoreMap[cr.Servers[j]] = true
@@ -95,9 +140,7 @@ func CronTrigger(cr model.Cron, triggerServer ...uint64) func() {
 			if len(triggerServer) == 0 {
 				return
 			}
-			ServerLock.RLock()
-			defer ServerLock.RUnlock()
-			if s, ok := ServerList[triggerServer[0]]; ok {
+			if s, ok := ServerShared.Get(triggerServer[0]); ok {
 				if s.TaskStream != nil {
 					s.TaskStream.Send(&pb.Task{
 						Id:   cr.ID,
@@ -108,15 +151,13 @@ func CronTrigger(cr model.Cron, triggerServer ...uint64) func() {
 					// 保存当前服务器状态信息
 					curServer := model.Server{}
 					copier.Copy(&curServer, s)
-					SendNotification(cr.NotificationTag, fmt.Sprintf("[任务失败] %s，服务器 %s 离线，无法执行。", cr.Name, s.Name), nil, &curServer)
+					NotificationShared.SendNotification(cr.NotificationGroupID, Localizer.Tf("[Task failed] %s: server %s is offline and cannot execute the task", cr.Name, s.Name), nil, &curServer)
 				}
 			}
 			return
 		}
 
-		ServerLock.RLock()
-		defer ServerLock.RUnlock()
-		for _, s := range ServerList {
+		for _, s := range ServerShared.Range {
 			if cr.Cover == model.CronCoverAll && crIgnoreMap[s.ID] {
 				continue
 			}
@@ -133,7 +174,7 @@ func CronTrigger(cr model.Cron, triggerServer ...uint64) func() {
 				// 保存当前服务器状态信息
 				curServer := model.Server{}
 				copier.Copy(&curServer, s)
-				SendNotification(cr.NotificationTag, fmt.Sprintf("[任务失败] %s，服务器 %s 离线，无法执行。", cr.Name, s.Name), nil, &curServer)
+				NotificationShared.SendNotification(cr.NotificationGroupID, Localizer.Tf("[Task failed] %s: server %s is offline and cannot execute the task", cr.Name, s.Name), nil, &curServer)
 			}
 		}
 	}
